% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking_functions.R
\name{add_mask_images}
\alias{add_mask_images}
\title{Add mask images}
\usage{
add_mask_images(r, ...)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Mask image, see \code{\link{mask_image}}.}

\item{...}{additional mask images.}
}
\value{
\linkS4class{RasterLayer}
}
\description{
Add mask images
}
\examples{
z <- zenith_image(1000, lens())
a <- azimuth_image(z)
m1 <- mask_image(z, zlim = c(20,70))
m2 <- mask_image(z, a, alim = c(180,350))
m <- add_mask_images(m1, m2)
}
\seealso{
write_bin

Other masking functions: 
\code{\link{mask_image}()}
}
\concept{masking functions}
