% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_thr.R
\name{apply_thr}
\alias{apply_thr}
\title{Apply a global threshold}
\usage{
apply_thr(r, thr)
}
\arguments{
\item{r}{\linkS4class{Raster}.}

\item{thr}{Numeric vector of length one. Threshold value.}
}
\value{
\linkS4class{RasterLayer} with values \code{0} and \code{1}.
}
\description{
Apply a global threshold value to an image.
}
\details{
It is a wrapper function around this operator from the 'raster' package:
\code{>}. The same threshold value is applied to every layer of the raster
object. The most frequent use is ingesting a single layered object, often
called a greyscale images.
}
\examples{
r <- read_caim()
apply_thr(r$Blue, 120)
\dontrun{
# This function is very useful in combination with the 'autothresholdr'
package. For examples:
thr <- autothresholdr::auto_thresh(r$Blue[], "IsoData")
apply_thr(r$Blue, thr[1])
}
}
