% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_sky_model.R
\name{autofit_cie_sky_model}
\alias{autofit_cie_sky_model}
\title{Autofit CIE sky model}
\usage{
autofit_cie_sky_model(r, z, a)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}
}
\description{
Autofit CIE sky model using data from real hemispherical photographs.
}
\details{
By using \code{\link{ootb_cie_mblt}}, it obtains a binarized image. Then use
it to extract the sky marks and sun coordinates required by
\code{\link{fit_CIE_sky_model}} (a sky grid of 10 degrees
--\code{sky_grid_segmentation(z, a, 10)}-- and the default parameters of
\code{\link{extract_sun_marks}} and \code{\link{extract_sun_mark}} are used).

However, the binarized image is not used as is, the \code{w} parameter (see
this concept on the \code{\link{thr_image}} documentation) is estimated based
on the sky border length calculated from the binarized image obtained with
\code{w} equal to \code{0.5} (by means of a Laplacian filter). The ratio of
the numer of pixels on the border to the total number of pixels is added to
\code{0.5}, but values above \code{0.9} are not allowed.

The argument \code{general_sky_type} of \code{\link{fit_CIE_sky_model}} is
obtained with \code{\link{choose_std_cie_sky_model}}.

If the general sky type is "Clear", since \code{\link{extract_sun_mark}} is
incapable of providing the sun position below the horizon, the sun below the
horizon (91 degrees) to the end of the astronomical twilight is explorer with
3 degrees resolution, and the best model is returned.

Since each general sky type has several standard CIE sky models, each of
those are tested and the best model is returned.

The best model is the one with low root mean square error.
}
\examples{
#' \dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
fit <- autofit_cie_sky_model(blue, z, a)
plot(fit$relative_luminance * fit$zenith_dn)
}
}
\seealso{
Other cie sky model functions: 
\code{\link{choose_std_cie_sky_model}()},
\code{\link{cie_sky_model_raster}()},
\code{\link{fit_cie_sky_model}()}
}
\concept{cie sky model functions}
