% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cone_shaped_model.R
\name{autofit_cone_shaped_model}
\alias{autofit_cone_shaped_model}
\title{Autofit cone shaped model}
\usage{
autofit_cone_shaped_model(r, z, a)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}
}
\description{
Out-of-the-box version of the \code{\link{fit_cone_shaped function}}
}
\examples{
#' \dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
sky <- autofit_cone_shaped_model(blue, z, a)
plot(sky$image)
}
}
