% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_image.R
\name{azimuth_image}
\alias{azimuth_image}
\title{Azimuth image}
\usage{
azimuth_image(z)
}
\arguments{
\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}
}
\value{
\code{\linkS4class{RasterLayer}}.
}
\description{
Built a single layer image with azimuth angles values.
}
\examples{
z <- zenith_image(1490, lens("Nikon_FCE9"))
azimuth_image(z)
plot(z)
}
\seealso{
zenith_image
}
