% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_co.R
\name{calc_co}
\alias{calc_co}
\title{Calculate canopy openness}
\usage{
calc_co(bin, z, a, m = NULL, angle_width = 10)
}
\arguments{
\item{bin}{\linkS4class{SpatRaster}. Binarized hemispherical canopy image.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{m}{\linkS4class{SpatRaster}. A mask. For hemispherical photographs,
check \code{\link[=select_sky_vault_region]{select_sky_vault_region()}}.}

\item{angle_width}{Numeric vector of length one. It should be \verb{30, 15, 10, 7.5, 6, 5, 3.75, 3, 2.5, 1.875, 1} or \code{0.5} degrees. This
constrain is rooted in the requirement of a value able to divide both the
\code{0} to \code{360} and \code{0} to \code{90} ranges into a whole number
of segments.}
}
\value{
Numeric vector of length one.
}
\description{
Calculate canopy openness
}
\details{
Canopy openness calculated as in the equation from
\insertCite{Gonsamo2011;textual}{rcaiman}:

\eqn{CO = \sum_{i = 1}^{N} GF(\phi_i, \theta_i) \cdot [(cos(\theta_1) -
cos(\theta_2))/n]},

where \eqn{GF(\phi_i, \theta_i)} is the gap fraction of the cell \eqn{i},
\eqn{\theta_1} and \eqn{\theta_2} are the minimum and maximum zenith angle of
the cell \eqn{i}, \eqn{n} is the number of cells on the ring delimited by
\eqn{\theta_1} and \eqn{\theta_2}, and \eqn{N} is the total number of cells.

When a mask is applied using the \code{m} argument, the equation is adjusted to
account for the reduced portion of the image that contributes to the
calculation:

\eqn{
\frac{
CO = \sum_{i = 1}^{N} GF(\phi_i, \theta_i) \cdot [(cos(\theta_1) -
cos(\theta_2))/n]
}{ \sum_{i = 1}^{N} (cos(\theta_1) - cos(\theta_2))/n}
}.

The denominator ensures that the canopy openness remains correctly scaled
even when part of the image is masked. Without this normalization, the total
openness could be underestimated since the summation in the numerator will be
one only when the full hemisphere of an totally unobstructed image is
included. Therefore, masking requires the normalization provided by the
denominator.
}
\examples{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)
m <- select_sky_vault_region(z, 0, 70)
bin <- apply_thr(caim$Blue, thr_isodata(caim$Blue[m]))
plot(bin)
calc_co(bin, z, a, m, 10)

}
\references{
\insertAllCited{}
}
\concept{Metrics Functions}
