% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_spherical_distance.R
\name{calc_spherical_distance}
\alias{calc_spherical_distance}
\title{Calculate spherical distance}
\usage{
calc_spherical_distance(z1, a1, z2, a2)
}
\arguments{
\item{z1}{Numeric vector. Zenithal angle in radians.}

\item{a1}{Numeric vector. Azimuthal angle in radians.}

\item{z2}{Numeric vector of length one. Zenithal angle in radians.}

\item{a2}{Numeric vector of length one. Azimuthal angle in radians.}
}
\value{
Numeric vector.
}
\description{
Calculate angles in radians between objects in the sky, which are
proportional to the shortest distances between then over a conceptual sky
vault.
}
\examples{
set.seed(1)
z1 <- rnorm(10, 45, 20) * pi/180
a1 <- rnorm(10, 180, 90) * pi/180
calc_spherical_distance(z1, a1, 0, 0)
}
