% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_functions.R
\name{calc_zenith_raster_coordinates}
\alias{calc_zenith_raster_coordinates}
\title{Calculate zenith raster coordinates}
\usage{
calc_zenith_raster_coordinates(path_to_csv)
}
\arguments{
\item{path_to_csv}{Character vector of length one. Path to a CSV file created
with ImageJ.}
}
\description{
Calculate zenith raster coordinates
}
\examples{
path <- system.file("external/points_over_perimeter.csv",
                    package = "rcaiman")
calc_zenith_raster_coordinates(path)
}
