% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_functions.R
\name{calibrate_lens}
\alias{calibrate_lens}
\title{Calibrate lens}
\usage{
calibrate_lens(path_to_csv, degree = 3)
}
\arguments{
\item{path_to_csv}{Character vector of length one. Path to a CSV file created
with ImageJ.}

\item{degree}{Numeric vector of length one. Polynomial model degree.}
}
\value{
named list.
}
\description{
Calibrate a fisheye lens.
}
\details{
TODO: describe how to obtain the csv
}
\examples{
path <- system.file("external/Results_calibration.csv", package = "rcaiman")
calibration <- calibrate_lens(path)
calibration$lens_coef
calibration$max_fov
calibration$max_fov_px
}
