% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_sky_model.R
\name{choose_std_cie_sky_model}
\alias{choose_std_cie_sky_model}
\title{Choose a standard CIE sky model}
\usage{
choose_std_cie_sky_model(
  r,
  z,
  a,
  bin,
  use_window = TRUE,
  twilight = TRUE,
  rmse = FALSE
)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{RasterLayer}. The result of binarizing the \code{r}
argument.}

\item{use_window}{Logical vector of length one. If \code{TRUE}, a 3 by 3
window will be used to extract the sky digital number from \code{r}.}

\item{twilight}{Logical vector of length one. If it is \code{TRUE} and the
initial standard parameters belong to the "Clear" general sky type, sun
zenith angles from 90 to 96 degrees will be tested (civic twilight). This
is necessary since \code{\link{extract_sun_marks}} would mistakenly
recognize the gravity center of what can be seen of the solar corona as the
solar disk.}

\item{rmse}{Logical vector of length one. If it is \code{TRUE}, the criteria
for selecting the best sky model is to choose the one with lest root mean
square error calculated from the sample (sky marks). Otherwise, the
criteria is to evaluate the whole hemisphere by calculating the ratio of
\code{r} to the sky model, then square it, and selecting the sky model that
produce the least value.}
}
\description{
Choose one from the 15th standard CIE sky models based on sky DN values.
}
\details{
The extraction of sky marks and sun coordinates is automatically done by the
function assuming the same that it is assumed for \code{\link{find_sky_dns}},
so please refer to that function help page, Details headline, first
paragraph.

Since a sky grid of 30 degrees is used, the maximum number of sky marks is
36, and the sun location is coarse. So, although it is not enough to estimate
model coefficients with \code{\link{fit_cie_sky_model}}, it can be used to
choose the closest standard sky.

To know more about the standard CIE sky models, please refer to
\insertCite{Li2016;textual}{rcaiman}.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
sky <- choose_std_cie_sky_model(blue, z, a)$relative_luminance
plot(sky)
}
}
\references{
\insertRef{Li2016}{rcaiman}
}
\seealso{
Other cie sky model functions: 
\code{\link{cie_sky_model_raster}()},
\code{\link{fit_cie_sky_model}()}
}
\concept{cie sky model functions}
