% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_sky_model.R
\name{cie_sky_model_raster}
\alias{cie_sky_model_raster}
\title{CIE sky model raster}
\usage{
cie_sky_model_raster(z, a, sun_coord, sky_coef)
}
\arguments{
\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}

\item{a}{\code{\linkS4class{RasterLayer}} built with
\code{\link{azimuth_image}}.}

\item{sun_coord}{Numeric vector of length 2. Zenith and azimuth angles in
degrees, corresponding to the location of the solar disk center.}

\item{sky_coef}{Numeric vector of length 5. Parameters of the sky model.}
}
\description{
CIE sky model raster
}
\examples{
\dontrun{
z <- zenith_image(1400, lens())
a <- azimuth_image(z)
path <- system.file("external", package = "rcaiman")
skies <- read.csv(file.path(path, "15_CIE_standard_skies.csv"))
# parameters are from http://dx.doi.org/10.1016/j.energy.2016.02.054
sky_coef <- skies[4,1:5]
sun_coord <- c(45, 0)
plot(cie_sky_model_raster(z, a, sun_coord, sky_coef))
}
}
\seealso{
Other cie sky model functions: 
\code{\link{fit_cie_sky_model}()}
}
\concept{cie sky model functions}
