% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_complementary_gradients.R
\name{compute_complementary_gradients}
\alias{compute_complementary_gradients}
\title{compute complementary gradients}
\usage{
compute_complementary_gradients(caim)
}
\arguments{
\item{caim}{\linkS4class{SpatRaster}. The output of \code{\link[=read_caim]{read_caim()}}.}
}
\value{
An object of class \linkS4class{SpatRaster}
}
\description{
compute complementary gradients
}
\examples{
\dontrun{
caim <- read_caim()
com <- compute_complementary_gradients(caim)
mem <- max(com$blue_yellow, com$cyan_red)

thrs <- calc_thrs(mem, z, a,  m,
                  angle_width = 30,
                  fov = 60,
                  method = "thr_twocorner_uc")

thrs_f <- filter_thrs(thrs, r, z, a, 2.5, FALSE, n_min = 0)

plot(apply_thr(mem, mean(thrs[,3])))
}
}
