% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defuzzify.R
\name{defuzzify}
\alias{defuzzify}
\title{Defuzzify fuzzy classification}
\usage{
defuzzify(mem, segmentation)
}
\arguments{
\item{mem}{An object of the class \linkS4class{SpatRaster}. Degree of
membership.}

\item{segmentation}{An object of the class \linkS4class{SpatRaster}, such as
the result of a call to \code{\link{sky_grid_segmentation}}.}
}
\value{
An object of the class \linkS4class{SpatRaster} containing binary
  information.
}
\description{
This function translates degree of membership into Boolean logic using an
regional approach. The result will ensure that at the chosen level of
aggregation (controlled by the argument \code{segmentation}) the fuzzy and
Boolean version will agree. This method is an alternative to applying a
global threshold with \code{\link{apply_thr}}.
}
\examples{
\dontrun{
path <- system.file("external/DSCN4500.JPG", package = "rcaiman")
r <- read_caim(path, c(1280, 960) - 745, 745 * 2, 745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
bin <- ootb_mblt(blue, z, a)
plot(bin$bin)
ratio <- blue / bin$sky_s
ratio[ratio > 1] <- 1
ratio[is.na(ratio)] <- 0
plot(ratio)
g <- sky_grid_segmentation(z, a, 10)
bin2 <- defuzzify(ratio, g)
}
}
