% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_sky_model.R
\name{.cie_sky_model}
\alias{.cie_sky_model}
\title{CIE sky model}
\usage{
.cie_sky_model(AzP, Zp, AzS, Zs, .a, .b, .c, .d, .e)
}
\arguments{
\item{AzP}{Numeric vector. Azimuth angle of a sky point.}

\item{Zp}{Numeric vector. Zenith Angle of a sky point.}

\item{AzS}{Numeric vector of length one. Azimuth angle of the sun.}

\item{Zs}{Numeric vector of length one. Zenith angle of the sun.}

\item{.a, .b, .c, .d, .e}{Numeric vector of length one. Sky model parameter.}
}
\value{
Numeric vector of length equal to AzP length.
}
\description{
Written by Gaston Mauro Diaz based on Pascal code by Mait Lang.
}
\details{
Angles should be provided in radians.
}
\references{
http://dx.doi.org/10.1016/j.energy.2016.02.054
}
\seealso{
Other cie sky model functions: 
\code{\link{cie_sky_model_raster}()},
\code{\link{fit_cie_sky_model}()}
}
\concept{cie sky model functions}
