% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_noncircular.R
\name{expand_noncircular}
\alias{expand_noncircular}
\title{Expand non-circular}
\usage{
expand_noncircular(caim, z, zenith_colrow)
}
\arguments{
\item{caim}{\linkS4class{RasterBrick}. The return of a call to
\code{\link{read_caim}}.}

\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}

\item{zenith_colrow}{Numeric vector of length two. Raster coordinates of the
zenith. See \code{\link{calc_zenith_raster_coordinates}}.}
}
\value{
caim \linkS4class{RasterBrick}
}
\description{
Expand a non-circular hemispherical photograph.
}
