% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_radiometry.R
\name{extract_radiometry}
\alias{extract_radiometry}
\title{Extract radiometry}
\usage{
extract_radiometry(path_to_pgms, z_thr = 15, size_px = NULL, pattern = "pgm")
}
\arguments{
\item{path_to_pgms}{Character vector of length one. Path to a folder with the
PGM files for radiometry sampling. These files must comply with the
equidistant projection.}

\item{z_thr}{Numeric vector of length one. Maximum angle in degrees used to
calculate the digital number (DN) not affected by the vignetting effect. If
\code{NULL} is provided, DNs are returned instead of relative luminance.}

\item{size_px}{Numeric vector of length one. Diameter in pixels of the
circular sampling area at the image center. This area is modified
considering the equidistant projection distortion. Therefore, it will be
visualized as an ellipse any other place but the image center.}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}
}
\value{
An object from the class \code{data.frame} with columns \emph{theta}
  (zenith angle in radians) and \emph{radiometry} (relative luminance or
  digital number, depending on argument \code{z_thr}.
}
\description{
Extract radiometry
}
\details{
Consult this
\href{https://docs.google.com/document/d/178yZDAcfx--Xn1Ye8Js-kUXuPCuYOHQL5fxAH7KBEoY/edit?usp=sharing}{document}
for additional details.

This code exemplify how to use the function to obtain the data and use base R
functions to obtain the vignetting function.
\preformatted{
up_data <- extract_radiometry("D610/up")
down_data <- extract_radiometry("D610/down")
right_data <- extract_radiometry("D610/right")
left_data <- extract_radiometry("D610/left")

ds <- rbind(up_data, down_data, right_data, left_data)

plot(ds, xlim = c(0, pi/2), ylim= c(0.5,1.05),
      col = c(rep(1,9),rep(2,9),rep(3,9),rep(4,9)))
legend("bottomleft", legend = c("up", "down", "right", "left"),
       col = 1:4, pch = 1)

fit <- lm((1 - ds$radiometry) ~ poly(ds$theta, 3, raw = TRUE) - 1)
summary(fit)
coef <- fit$coefficients
fun <- function(x) 1 - (coef[1] * x + coef[2] * x^2 + coef[3] * x^3)
curve(fun, add = TRUE, col= 2)
coef
}
}
\seealso{
Other Lens Functions: 
\code{\link{azimuth_image}()},
\code{\link{calc_diameter}()},
\code{\link{calc_zenith_raster_coord}()},
\code{\link{calibrate_lens}()},
\code{\link{expand_noncircular}()},
\code{\link{fisheye_to_equidistant}()},
\code{\link{fisheye_to_pano}()},
\code{\link{lens}()},
\code{\link{test_lens_coef}()},
\code{\link{zenith_image}()}
}
\concept{Lens Functions}
