% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rl.R
\name{extract_rl}
\alias{extract_rl}
\title{Extract relative luminance}
\usage{
extract_rl(
  r,
  z,
  a,
  sky_points,
  no_of_points = 20,
  z_thr = 2,
  use_window = TRUE
)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photograph. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{azimuth_image}}.}

\item{sky_points}{An object of class data.frame. The result of a call to
\code{\link{extract_sky_points}}.}

\item{no_of_points}{Numeric vector on length one. The number of near zenith
points required for the estimation of the zenith DN.}

\item{z_thr}{Numeric vector on length one. The starting maximum zenith angle
used to search for near zenith points.}

\item{use_window}{Logical vector of length one. If \code{TRUE}, a 3 by 3
window will be used to extract the sky digital number from \code{r}.}
}
\value{
A list of three objects, ‘zenith_dn’ and ‘max_zenith_angle’ from the
  class numeric, and ‘sky_points’ from the class data.frame; ‘zenith_dn’ is
  the estimated zenith digital number, ‘max_zenith_angle’ is the maximum
  azimuth reached in the search for near zenith sky points, and ‘sky_points’
  is the input argument ‘sky_points’ with the additional columns: \emph{a},
  \emph{z}, and \emph{dn}, and \emph{rl}, which stand for azimuth and zenith
  angle in degrees, digital number, and relative luminance, respectively. If
  \code{NULL} is provided as \code{no_of_points}, then ‘zenith_dn’ is forced
  to one and \emph{dn}, and \emph{rl} are equals.
}
\description{
Extract the luminance relative to the zenith digital number (zenith DN).
}
\details{
To estimate the zenith DN, the search for near zenith points starts in the
region delimited by \code{z_thr}. If the number of near zenith points is less
than \code{no_of_points}, the region increase by 2 degrees of zenith angle
til the required number of points is reached
}
\examples{
\dontrun{
path <- system.file("external/DSCN4500.JPG", package = "rcaiman")
r <- read_caim(path, c(1280, 960) - 745, 745 * 2, 745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
g <- sky_grid_segmentation(z, a, 10)
bin <- find_sky_pixels(blue, z, a)$bin
sky_points <- extract_sky_points(blue, bin, g)
rl <- extract_rl(blue, z, a, sky_points, 1)
}
}
\seealso{
Other MBLT functions: 
\code{\link{extract_sky_points}()},
\code{\link{find_sky_pixels}()},
\code{\link{fit_coneshaped_model}()},
\code{\link{fit_trend_surface}()},
\code{\link{ootb_mblt}()},
\code{\link{thr_image}()}
}
\concept{MBLT functions}
