% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sun_mark.R
\name{extract_sun_coord}
\alias{extract_sun_coord}
\title{Extract sun coordinates}
\usage{
extract_sun_coord(r, z, a, bin, g, max_angular_dist = 30)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. Greyscale image hemispherical image.}

\item{z}{\linkS4class{SpatRaster} built with
\code{\link{zenith_image}}.}

\item{a}{\code{\linkS4class{SpatRaster}} built with
\code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{SpatRaster}. The result of binarizing the \code{r}
argument.}

\item{g}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{sky_grid_segmentation}} taking into account the camera, lens,
and pre-processing involved in obtaining the \code{r} argument.}

\item{max_angular_dist}{Numeric vector of length one. Angle in degree to
establish the maximum size of the sun corona. See details. The former is
raster coordinates of the solar disk (row and column), and the other is
angular coordinates (zenith and azimuth angles in degrees).}
}
\value{
Object of class list with two elements names "row_col" and
  "zenith_azimuth", both are numeric vectors of length two.
}
\description{
Extract the sun coordinates for CIE sky model fitting.
}
\details{
This function use an object-based image analyze theoretical framework. The
segmentation are given by \code{g} and \code{bin}. For every cell of
\code{g}, the maximum is calculated from the pixel values on \code{r} that
have a value equal to one on \code{bin}. Then, the quantile 0.95 is extracted
from these maximum values and is used to filter out cells below that
threshold, i.e, only the ones with at least one extremely bright sky pixel is
keep.

Selected cells are grouped into segments based on adjacency. The degree of
membership to the class _Sun_ is calculated for every segment by using linear
membership functions for the features brightness --digital number from
\code{r}-- and size -- number of cells that constitute the segment. In other
words, the brighteners and lagers segments are the ones that score higher.
The one with the highest score is selected as a sun seed.

The angular distance from the sun seed to every other segments are computed,
and only the segments not farther than \code{max_angular_dist} are classified
as part of the sun corona. A multi-part segment is created by merging the
sun-corona segments, the center of its bounding box is considered as the sun
location.

The \code{bin} argument should be the same than for
\code{\link{extract_sky_points}}.
}
\examples{
\dontrun{
caim <- read_caim() \%>\% normalize()
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
sky_blue_sample <- crop(caim, ext(610,643,760,806))
sky_blue <- apply(sky_blue_sample[], 2, median) \%>\%
  as.numeric() \%>\%
  matrix(., ncol = 3) \%>\%
  sRGB()
ecaim <- enhance_caim(caim, !is.na(z), sky_blue, gamma = 2.2)
bin <- apply_thr(ecaim, 0.75)
g <- sky_grid_segmentation(z, a, 10)
blue <- gbc(caim$Blue*255)
sun_coord <- extract_sun_coord(blue, z, a, bin, g, max_angular_dist = 30)
xy <- cellFromRowCol(z, sun_mark$row_col[1], sun_mark$row_col[2]) \%>\%
  xyFromCell(z, .)
plot(blue)
plot(vect(xy), add = TRUE, col = 2)
}
}
\concept{hsp functions}
