% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{extract_sun_mark}
\alias{extract_sun_mark}
\title{Extract sun mark}
\usage{
extract_sun_mark(r, bin, g, path_to_HPS_project, img_name)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Greyscale image hemispherical image.}

\item{bin}{\linkS4class{RasterLayer}. The result of binarizing the \code{r}
argument.}

\item{g}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{sky_grid_segmentation}} taking into account the camera, lens,
and pre-processing involved in obtaining the \code{r} argument.}

\item{path_to_HPS_project}{Character vector of length one. Path to the HPS
project folder.}

\item{img_name}{Character vector of length one. The name of the file
associated to the \code{r} argument.}
}
\description{
Extract sun mark and create a special file to interface with HSP software.
}
\seealso{
extract_sky_marks
}
