% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_points.R
\name{fibonacci_points}
\alias{fibonacci_points}
\title{Map Fibonacci points to raster coordinates}
\usage{
fibonacci_points(z, a, spacing)
}
\arguments{
\item{z}{\link[terra:SpatRaster-class]{terra::SpatRaster} generated with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\link[terra:SpatRaster-class]{terra::SpatRaster} generated with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{spacing}{numeric vector of length one. Angular separation (deg) between}
}
\value{
\code{data.frame} with integer columns \code{row} and \code{col}.
}
\description{
Return image row and column indices for points distributed on the upper
hemisphere using a spherical Fibonacci lattice with approximately constant
angular spacing.
}
\examples{
z <- zenith_image(100, lens())
a <- azimuth_image(z)
sampling_points <- fibonacci_points(z, a, 30)
\dontrun{
display_caim(is.na(z), sky_points = sky_points)
}
}
\seealso{
\code{\link[=skygrid_centers]{skygrid_centers()}}
}
