% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_thrs.R
\name{filter_thrs}
\alias{filter_thrs}
\title{Title}
\usage{
filter_thrs(
  thrs,
  r,
  z,
  a,
  g = NULL,
  laxity = 2.5,
  per_direction = TRUE,
  n_min = 20
)
}
\arguments{
\item{thrs}{Character vector of length one.}

\item{r}{\linkS4class{SpatRaster}. The image from which the argument \code{sky_points}
was obtained.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{laxity}{Numeric vector of length one. See
\insertCite{Leys2013;textual}{rcaiman}.}
}
\value{
a list.
}
\description{
Title
}
\examples{
\dontrun{
}
}
