% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_surface_to_sky_dn.R
\name{fit_surface_to_sky_dn}
\alias{fit_surface_to_sky_dn}
\title{Fit a trend surface to sky digital numbers}
\usage{
fit_surface_to_sky_dn(
  x,
  z,
  mask,
  bin,
  filling_source = NULL,
  prob = 0.95,
  fact = 5,
  np = 6
)
}
\arguments{
\item{x}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{mask}{raster}

\item{bin}{\linkS4class{RasterLayer}. A working binarized image. This should
be a preliminary binarization of \code{x}. If the function returns
\code{NA}, the quality of this input should be revised.}

\item{filling_source}{\linkS4class{RasterLayer}. Default is \code{NULL}.
Above-canopy photograph. This image should contain pixels with sky DN
values and \code{NA} in all the other pixels. A photograph taken
immediately after or before taking \code{x} under the open sky with the
same equipment and configuration is a very good option. The ideal option is
one taken at the same time and place but above the canopy. The orientation
relative to north must be the same than \code{x}.}

\item{prob}{One-length logical vector. Probability for
\code{\link[stats]{quantile}}. See reference
\insertCite{Diaz2018;textual}{rcaiman}.}

\item{fact}{postive integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). See Details}

\item{np}{
degree of polynomial surface
}
}
\value{
\code{\linkS4class{RasterLayer}}
}
\description{
Fit a trend surface using spatial::surf.ls as workhorse function.
}
\details{
This function is meant to be used before \code{\link{model_sky_dn}}.

A short explanation of this function can be found on
\insertCite{Diaz2018;textual}{rcaiman}, under the heading \emph{Estimation of
the sky DN as a previous step for our method}, after the explanation of the
\code{\link{model_sky_dn}}.

The argument \code{fact} is passed to the \code{\link[raster]{aggregate}}.
Whit this parameter, it is possible to control the scale at which the fitting
is performed. In general, a coarse scale lead to best generalization.

If a stake of above canopy references is provided as filling source, the
function can select the one with less root mean square by comparing them on
the areas were sky DN are available in both \code{x} and
\code{filling_source}.
}
\examples{
a <- 10
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
mblt functions
}
