% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend_surface_to_sky_dn.R
\name{fit_trend_surface_to_sky_dn}
\alias{fit_trend_surface_to_sky_dn}
\title{Fit a trend surface to sky digital numbers}
\usage{
fit_trend_surface_to_sky_dn(
  r,
  z,
  mask,
  bin,
  filling_source = NULL,
  prob = 0.95,
  fact = 5,
  np = 6
)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{mask}{\linkS4class{RasterLayer}. Usually, the result of a call to
\code{\link{mask_image}}.}

\item{bin}{\linkS4class{RasterLayer}. A working binarized image. This should
be a preliminary binarization of \code{x}. If the function returns
\code{NA}, the quality of this input should be revised.}

\item{filling_source}{\linkS4class{RasterLayer}. Default is \code{NULL}.
Above-canopy photograph. This image should contain pixels with sky DN
values and \code{NA} in all the other pixels. A photograph taken
immediately after or before taking \code{x} under the open sky with the
same equipment and configuration is a very good option. The ideal option is
one taken at the same time and place but above the canopy. The orientation
relative to north must be the same than \code{x}.}

\item{prob}{One-length logical vector. Probability for
\code{\link[stats]{quantile}} calculation. See reference
\insertCite{Diaz2018;textual}{rcaiman}.}

\item{fact}{postive integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). See Details}

\item{np}{
degree of polynomial surface
}
}
\value{
A list with an object of class \linkS4class{RasterLayer} and of class
  "trls" (see \code{\link[spatial]{surf.ls}}).
}
\description{
Fit a trend surface using spatial::surf.ls as workhorse function.
}
\details{
This function is meant to be used after \code{\link{model_sky_dn}}.

A short explanation of this function can be found on
\insertCite{Diaz2018;textual}{rcaiman}, under the heading \emph{Estimation of
the sky DN as a previous step for our method}, after the explanation of the
\code{\link{model_sky_dn}}.

The argument \code{fact} is passed to \code{\link[raster]{aggregate}}. That
argument allows to control the scale at which the fitting is performed. In
general, a coarse scale lead to best generalization.
}
\examples{
\dontrun{
path <- getwd()
my_file <- paste0(path, "/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
  method = "auto", mode = "wb"
)
r <- read_caim(
  my_file,
  c(1280, 960) - 745,
  745 * 2,
  745 * 2
)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
thr <- autothresholdr::auto_thresh(r$Blue[], "IsoData")
bin <- apply_thr(r$Blue, thr[1] * 1.25)
blue <- gbc(r$Blue)
sky <- model_sky_dn(blue, z, a, bin, parallel = FALSE)
m <- mask_image(z, zlim = c(0, 70))
sky <- fit_trend_surface_to_sky_dn(blue, z, m, bin,
  filling_source = sky$image
)
}
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
Other mblt functions: 
\code{\link{mblt}()},
\code{\link{model_sky_dn}()},
\code{\link{thr_image}()}
}
\concept{mblt functions}
