% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{gbc}
\alias{gbc}
\title{Gamma back correction}
\usage{
gbc(DN_from_JPEG, gamma = 2.2)
}
\arguments{
\item{DN_from_JPEG}{Numeric vector or an object from the
\code{\linkS4class{Raster}} class. Digital numbers from a JPEG file.}

\item{gamma}{Numeric vector of length one. Gamma value. Please see the
reference for details.}
}
\value{
Normalized \code{\linkS4class{Raster}}.
}
\description{
Gamma back correction of JPEG images.
}
\examples{
r <- read_caim()
gbc(r)
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
\code{\link{normalize}}

Other Tools functions: 
\code{\link{apply_thr}()},
\code{\link{extract_feature}()},
\code{\link{masking}()},
\code{\link{normalize}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{regional_thresholding}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()}
}
\concept{Tools functions}
