% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate_as_is}
\alias{interpolate_as_is}
\title{Interpolate as is}
\usage{
interpolate_as_is(r, sky_marks, k = 3, p = 2, rmax = 200, use_window = TRUE)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{sky_marks}{An object of class data.frame. The result of a call to
\code{\link{extract_sky_marks}}.}

\item{k}{Integer vector of length one. Number of k-nearest neighbors.}

\item{p}{Numeric vector of length one. Power for inverse-distance weighting.}

\item{rmax}{Numeric vector of length one. Maximum radius where to search for
knn.}

\item{use_window}{Logical vector of length one. If \code{TRUE}, a 3 by 3
window will be used to extract the sky digital number from \code{r}.}
}
\description{
Interpolate values from hemispherical photographs as is.
}
\details{
This function is based on \insertCite{Lang2010;textual}{rcaiman}. In theory,
interpolation requires a linear relation between DNs and the amount of light
reaching the sensor. To that end, photographs should be taken in RAW format
to avoid gamma correction \insertCite{Lang2010}{rcaiman}. As a compromise
solution, \code{\link{gbc}} can be used.

The vignetting effect also hindered linear relation between DNs and the
amount of light reaching the sensor. Please refer to
\insertCite{Lang2010;textual}{rcaiman} for more details about the vignetting
effect.

The use of \code{k = 1} solve the linear dilemma from the theoretical point
of view since no averaging is taking place in the calculations.

Default parameters are the used by \insertCite{Lang2010;textual}{rcaiman}.
The argument \code{rmax} should account for between 15 to 20 degrees, but is
expressed in pixels units. So, image resolution and lens projections should
be taken into account to properly set this parameter.

This function use \code{\link[lidR]{knnidw}} as workhorse function.
}
\examples{
#' \dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
bin <- ootb_mblt(blue, z, a, is_horizon_visible = TRUE)$bin
g <- sky_grid_segmentation(z, a, 10)
sky_marks <- extract_sky_marks(blue, bin, g)
sky <- interpolate_as_is(blue, sky_marks)
plot(sky)
}
}
\references{
\insertRef{Lang2010}{rcaiman}
}
\seealso{
interpolate_reproj
}
