% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_dns.R
\name{interpolate_dns}
\alias{interpolate_dns}
\title{Interpolate DNs}
\usage{
interpolate_dns(
  r,
  z,
  a,
  sky_marks,
  lens_coef,
  k = 3,
  p = 2,
  rmax = 20,
  use_window = TRUE
)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}

\item{sky_marks}{An object of class data.frame. The result of a call to
\code{\link{extract_sky_marks}}.}

\item{lens_coef}{Numeric vector. Polynomial coefficients
of the lens projection function.}

\item{k}{integer. Number of k-nearest neighbours. Default 10.}

\item{p}{numeric. Power for inverse-distance weighting. Default 2.}

\item{rmax}{numeric. Maximum radius where to search for knn. Default 50.}

\item{use_window}{Logical vector of length one. If \code{TRUE}, a 3 by 3
window will be used to extract the sky digital number from \code{r}.}
}
\description{
Interpolate digital numbers using lidR::knnidw as workhorse function.
}
\details{
This function is based on \insertCite{Lang2010;textual}{rcaiman}. In theory,
interpolation requires a linear relation between DNs and the amount of light
reaching the sensor. The photographs should be taken in RAW format to avoid
gamma correction \insertCite{Lang2010}{rcaiman}. As a compromise solution,
\code{\link{gbc}} can be used.

The vignetting effect also hindered linear relation between DNs and the
amount of light reaching the sensor. Please refer to
\insertCite{Lang2010;textual}{rcaiman} for more details about the vignetting
effect.

The use of \code{p = 1} solve the linear dilemma from the theoretical point
of view since no averaging is taking place in the calculations.

This function assume an hemispherical image as input. The interpolation takes
place after a cylindrical reprojection is performed. Thirty-degree portions
of the reprojected image are taken from the margin and duplicated to ensure
hemispherical continuity. The reprojected image has resolution of one degree.
Thus, \code{rmax} is in degrees.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
bin <- ootb_mblt(blue, z, a, is_horizon_visible = TRUE)$bin
sky_marks <- extract_sky_marks(blue, bin, g)
sky <- interpolate_dns(blue, z, a, sky_marks, lens("Nikon_FCE9"))
plot(sky)
}
}
\references{
\insertRef{Lang2010}{rcaiman}
}
