% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fuzzy_thresholding.R
\name{local_fuzzy_thresholding}
\alias{local_fuzzy_thresholding}
\title{Local fuzzy thresholding}
\usage{
local_fuzzy_thresholding(lightness, m, mem, thr = NULL, fuzziness = NULL)
}
\arguments{
\item{lightness}{\linkS4class{SpatRaster}. A normalized greyscale image (see
\code{\link[=normalize_minmax]{normalize_minmax()}}).}

\item{m}{\linkS4class{SpatRaster}. A mask. For hemispherical photographs,
check \code{\link[=select_sky_vault_region]{select_sky_vault_region()}}.}

\item{mem}{\linkS4class{SpatRaster}. It is the scale parameter of the logistic
membership function. Typically it is obtained with \code{\link[=membership_to_color]{membership_to_color()}}.}

\item{thr}{Numeric vector of length one. Location parameter of the logistic
membership function. Use \code{NULL} to estimate it automatically with
\code{\link[=thr_isodata]{thr_isodata()}}.}

\item{fuzziness}{Numeric vector of length one. This number is a constant
value that scales \code{mem}. Use \code{NULL} to estimate it automatically as the
midpoint between the maximum and minimum values of \code{lightness}.}
}
\value{
An object of class \linkS4class{SpatRaster} with same pixel dimensions than
\code{caim}. Depending on \code{mem}, changes could be subtle.
}
\description{
This function was first presented in \insertCite{Diaz2015;textual}{rcaiman}.
It uses a threshold value as the location parameter of a logistic membership
function whose scale parameter depends on a variable, here named \code{mem}. This
dependence can be explained as follows: if the variable is equal to \code{1}, then
the membership function is same as a threshold function because the scale
parameter is \code{0}; lowering the variable increases the scale parameter, thus
blurring the threshold because it decreases the steepness of the curve. Since
the variable is defined pixel by pixel, this should be considered as a
\strong{local} fuzzy thresholding method.
}
\details{
Argument \code{m} can be used to affect the automatic estimation of \code{thr} and
\code{fuzziness}.

If you use this function in your research, please cite
\insertCite{Diaz2015;textual}{rcaiman} in addition to this package
(\verb{citation("rcaiman"}).
}
\examples{
\dontrun{
path <- system.file("external/APC_0581.jpg", package = "rcaiman")
caim <- read_caim(path)
z <- zenith_image(2132/2, lens("Olloclip"))
a <- azimuth_image(z, orientation = 180)
zenith_colrow <- c(1063, 771)/2
caim <- expand_noncircular(caim, z, zenith_colrow)
m <- !is.na(caim$Red) & !is.na(z)
caim[!m] <- 0
caim <- normalize_minmax(caim)

mem <- membership_to_color(caim, sRGB(0.1, 0.4, 0.8))
plot(mem)
mem_thr <- local_fuzzy_thresholding(caim$Blue, m,
                                    mem$membership_to_grey,
                                    fuzziness = 0.1)
plot(mem_thr)
}
}
\references{
\insertAllCited{}
}
