% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking_functions.R
\name{mask_hemisphere}
\alias{mask_hemisphere}
\title{Mask hemisphere}
\usage{
mask_hemisphere(r, from, to)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}} or \code{\link{azimuth_image}}}

\item{from, to}{angle in degrees, inclusive limits.}
}
\description{
Given a zenith or azimuth image and angle restrictions, it produces a mask.
}
\examples{
z <- zenith_image(1000, lens())
a <- azimuth_image(z)
m1 <- mask_hemisphere(z, 20, 70)
plot(m1)
m2 <- mask_hemisphere(a, 330,360)
plot(m2)
plot(m1 & m2)
plot(m1 | m2)

# if you want 15 degress at each side of 0
m1 <- mask_hemisphere(a, 0, 15)
m2 <- mask_hemisphere(a, 345, 360)
plot(m1 | m2)

# better use this
plot(!is.na(z))
# instead of this
plot(mask_hemisphere(z, 0, 90))
}
