% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking_functions.R
\name{mask_image}
\alias{mask_image}
\title{Mask image}
\usage{
mask_image(z, a = NULL, zlim = NULL, alim = NULL, previous_mask = NULL)
}
\arguments{
\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}

\item{a}{\code{\linkS4class{RasterLayer}} built with
\code{\link{azimuth_image}}.}

\item{zlim}{Numeric vector of length two. Angles in degrees. Set the zenith
angle range with inclusive limits.}

\item{alim}{Numeric vector of length two. Angles in degrees. Set the azimuth
angle range with inclusive limits.}

\item{previous_mask}{\linkS4class{RasterLayer}. Typically, a mask extracted
from an expanded non-circular hemispherical photograph (see
\code{\link{expand_noncircular}}).}
}
\value{
\linkS4class{RasterLayer}
}
\description{
Given angle restrictions, produce an image for masking fisheye photos.
}
\examples{
z <- zenith_image(1000, lens())
a <- azimuth_image(z)

m <- mask_image(z, a, c(20, 70), c(90, 180))
plot(m)

m <- mask_image(z, zlim = c(20, 70))
plot(m)

m <- mask_image(z)
plot(m)
}
\seealso{
\code{\link{write_bin}}

Other masking functions: 
\code{\link{add_mask_images}()}
}
\concept{masking functions}
