% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking_functions.R
\name{masking}
\alias{masking}
\alias{masking,RasterLayer-method}
\alias{masking,RasterStackBrick-method}
\title{Image masking}
\usage{
masking(r, m, RGB = c(1, 0, 0))

\S4method{masking}{RasterLayer}(r, m, RGB = c(1, 0, 0))

\S4method{masking}{RasterStackBrick}(r, m, RGB = c(1, 0, 0))
}
\arguments{
\item{r}{\linkS4class{Raster}. The image. Values should be normalized, see
\code{\link{normalize}}.}

\item{m}{\linkS4class{RasterLayer}. The mask, see
\code{\link{mask_hemisphere}}.}

\item{RGB}{Numeric vector of length three. RGB color code. Red is the default
color.}
}
\value{
\linkS4class{RasterStack}
}
\description{
Image masking
}
\examples{
 r <- read_caim()
 z <- zenith_image(ncol(r), lens())
 a <- azimuth_image(z)
 m <- mask_hemisphere(z, 20, 70) & mask_hemisphere(a, 90, 180)

 masked_caim <-  masking(normalize(r, 0, 255), m)
 plotRGB(masked_caim * 255)

 masked_bin <- masking(apply_thr(r$Blue, 125), m)
 plotRGB(masked_bin * 255)
}
