% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblt_subpixel.R
\name{mblt_subpixel}
\alias{mblt_subpixel}
\title{MBLT subpixel}
\usage{
mblt_subpixel(r, m, bin_plant, bin_sky, parallel = TRUE, free_threads = 2)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{bin_plant}{\linkS4class{RasterLayer}. A binarization of r that works
as a mask for pure plant pixels.}

\item{bin_sky}{\linkS4class{RasterLayer}. A binarization of r that works
as a mask for pure sky pixels.}

\item{parallel}{One-length logical vector. Allows parallel processing.}
}
\value{
\linkS4class{RasterLayer}
}
\description{
MBLT subpixel
}
\examples{
\dontrun{
path <- getwd()
my_file <- paste0(path, "/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
bin <- mblt(blue, z, a, w = c(0.1,0.9,0.5))
m <- mask_image(z, zlim = c(0,70))
bin2 <- mblt_subpixel(blue, m,
                      bin_plant = !bin$w0.1,
                      bin_sky = bin$w0.9)
}
}
