% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_caim.R
\name{membership_to_color}
\alias{membership_to_color}
\title{Compute membership to a color}
\usage{
membership_to_color(caim, target_color, sigma = NULL)
}
\arguments{
\item{caim}{\linkS4class{RasterBrick}. The return of a call to
\code{\link{read_caim}}.}

\item{target_color}{\linkS4class{color}.}

\item{sigma}{Numeric vector of length one. Use \code{NULL} (default) to order
to estimate it automatically as the euclidean distance between
\code{target_color} and grey in the \emph{CIE L*a*b*} color space.}
}
\value{
It returns an object from the class \linkS4class{RasterBrick} or
  \linkS4class{RasterStack} --this will depend on the input. First layer is
  the membership to the target color. Second layer is the membership to grey.
  Both membership are calculated with same \code{sigma}.
}
\description{
Compute the degree of membership to a color with two Gaussian membership
functions and the dimensions \emph{A} and \emph{B} from the \emph{CIE L*a*b*}
color space. The lightness dimension is not considered in the calculations.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSC_2881.JPG")
download.file("https://osf.io/x8urg/download", my_file,
              method = "auto", mode = "wb"
)

caim <- read_caim(my_file)
caim <- normalize(caim, 0, 255)
target_color <- sRGB(matrix(c(0.529, 0.808, 0.921), ncol = 3))
mem <- membership_to_color(caim, target_color)
plot(mem)
}
}
