% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sky_dn.R
\name{model_sky_dn}
\alias{model_sky_dn}
\title{Model sky digital numbers}
\usage{
model_sky_dn(
  x,
  z,
  a,
  bin,
  prob = 0.95,
  filling_source = NULL,
  use_azimuth_angle = TRUE,
  parallel = TRUE,
  free_cores = 0
)
}
\arguments{
\item{x}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{RasterLayer}. A working binarized image. This should
be a preliminary binarization of \code{x}. If the function returns
\code{NA}, the quality of this input should be revised.}

\item{prob}{One-length logical vector. Probability for
\code{\link[stats]{quantile}}. See reference
\insertCite{Diaz2018;textual}{rcaiman}.}

\item{filling_source}{\linkS4class{RasterLayer}. Default is \code{NULL}.
Above-canopy photograph. This image should contain pixels with sky DN
values and \code{NA} in all the other pixels. A photograph taken
immediately after or before taking \code{x} under the open sky with the
same equipment and configuration is a very good option. The ideal option is
one taken at the same time and place but above the canopy. The orientation
relative to north must be the same than \code{x}.}

\item{use_azimuth_angle}{One-length logical vector. If \code{TRUE}, Equation
4 from \insertCite{Diaz2018;textual}{rcaiman} is used: \eqn{sDN = a + b
\cdot \theta + c  \cdot \theta^2 + d  \cdot sin(\phi) + e  \cdot
cos(\phi)}, where \eqn{sDN} is sky digital number, \eqn{a,b,c,d} and
\eqn{e} are coefficients, \eqn{\theta} is zenith angle, and \eqn{\phi} is
azimuth angle. If \code{FALSE}, a simplified version based on
\insertCite{Wagner2001;textual}{rcaiman} is used: \eqn{sDN = a + b \cdot
\theta + c  \cdot \theta^2}.}

\item{parallel}{One-length logical vector. Allows parallel processing.}

\item{free_cores}{One-length numeric vector. This number is subtracted to the
number of cores detected by \code{\link[parallel]{detectCores}}.}
}
\value{
a list with an object of class \linkS4class{RasterLayer} and of
  class "lm" (see \code{\link[stats]{lm}}).
}
\description{
Produce the digital numbers of the whole sky through statistical modelling.
}
\details{
An explanation of this function can be found on
\insertCite{Diaz2018;textual}{rcaiman}, under the heading \emph{Estimation of
the sky DN as a previous step for our method}.
}
\examples{
\dontrun{
path <- getwd()
my_file <- paste0(path, "/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
               method = "auto", mode = "wb"
)

r <- read_caim(file.path(path, "DSCN5548.JPG"),
               c(1280, 960) - 745, 745 * 2, 745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
thr <- autothresholdr::auto_thresh(r$Blue[], "IsoData")
bin <- apply_thr(r$Blue, thr[1] * 1.25)
blue <- gbc(r$Blue)
sky <- model_sky_dn(blue, z, a, bin, parallel = FALSE)
path <- tempfile(fileext = ".tif")
write_caim(sky$image * 2^8, path, 8)
}
}
\references{
\insertRef{Diaz2018}{rcaiman}

\insertRef{Wagner2001}{rcaiman}
}
\seealso{
Other mblt functions: 
\code{\link{fit_surface_to_sky_dn}()}
}
\concept{mblt functions}
