% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obia.R
\name{obia}
\alias{obia}
\title{Object-based image analysis of canopy photographs}
\usage{
obia(r, z = NULL, a = NULL, bin, segmentation, gf_mn = 0.2, gf_mx = 0.95)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically,
the blue channel extracted from a canopy photograph. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link{zenith_image}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{SpatRaster}. This should be a working binarization
of \code{r} without gross errors.}

\item{segmentation}{\linkS4class{SpatRaster} built with
\code{\link{polar_qtree}}.}

\item{gf_mn, gf_mx}{Numeric vector of length one. The minimum/maximum gap
fraction that a segment should comply with to be considered as one
containing foliage.}
}
\value{
\linkS4class{SpatRaster}.
}
\description{
This method was first presented in \insertCite{Diaz2015;textual}{rcaiman}.
This version is simpler since it relies on a better working binarized image.
The version from 2015 uses an automatic selection of samples followed by a
\emph{knn} classification of segments containing foliage. This version uses
de gap fraction extracted from \code{bin} to classify \emph{foliage} by
defining upper and lower limits through the arguments \code{gf_mx} and
\code{gf_mn}.
}
\details{
This method produces a synthetic layer by computing the ratio of \code{r} to
the maximum value of \code{r} at the segment level. This process is carried
out only on the pixels covered by the classes \emph{foliage} and \emph{sky}--
the latter is defined by bin equal to one. To avoid spurious values, the
quantile \code{0.9} is computed instead of the maximum. Pixels not belonging
to the class \emph{foliage} return as \code{NA}.
}
\examples{
\dontrun{
#hemispherical photo from a smartphone
caim <- read_caim()
r <- caim$Blue \%>\% gbc()
caim <- normalize(caim)
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
seg <- polar_qtree(caim, z, a)
ecaim <- enhance_caim(caim, !is.na(z))
bin <- apply_thr(ecaim, thr_isodata(ecaim[]))
synth <- obia(r, z, a, bin, seg)
plot(synth)
synth <- terra::cover(synth, bin)
plot(synth)
synth[!bin] <- 0
plot(synth)
bin_obia <- apply_thr(synth, thr_isodata(synth[]))
plot(bin - bin_obia)
plot(bin_obia)
## This is the closest to the workflow presented in Diaz and Lencinas
## (2015) that this package allows.

#restricted view canopy photo
path <- system.file("external/APC_0020.jpg", package = "rcaiman")
caim <- read_caim(path)
plot(caim)
blue <- gbc(caim$Blue)
caim <- normalize(caim)

ecaim <- enhance_caim(caim, thr = 1, fuzziness = 1)
plot(ecaim)
bin <- apply_thr(ecaim, thr_isodata(ecaim[]))

seg <- qtree(caim)
synth <- obia(blue, NULL, NULL, bin, seg)
plot(synth)
synth <- terra::cover(synth, bin)
plot(synth)
synth[!bin] <- 0
plot(synth)
bin_obia <- apply_thr(synth, thr_isodata(synth[]))
plot(bin - bin_obia)
plot(bin_obia)
}
}
\references{
\insertAllCited{}
}
\seealso{
Other Binarization Functions: 
\code{\link{apply_thr}()},
\code{\link{find_sky_pixels_nonnull}()},
\code{\link{find_sky_pixels}()},
\code{\link{ootb_mblt}()},
\code{\link{ootb_obia}()},
\code{\link{regional_thresholding}()},
\code{\link{thr_image}()},
\code{\link{thr_isodata}()}
}
\concept{Binarization Functions}
