% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblt.R
\name{ootb_cie_mblt}
\alias{ootb_cie_mblt}
\title{Out-of-the-box model-based local thresholding with CIE sky model included}
\usage{
ootb_cie_mblt(r, z, a)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}
}
\value{
Object of class list with the binarized image (named "bin") and the
  reconstructed skies named as follow: "sky_cs" is the cone shaped model,
  "std_cie_sky" is the output from \code{\link{choose_std_cie_sky_model}},
  "sky_s" is the trend surface, and "sky" is the actually used to obtain
  "bin" (please see Details).
}
\description{
The same as \code{\link{ootb_mblt}}, it is a hard-coded version of a MBLT
pipeline. The code can be easily inspected by calling \code{ootb_cie_mblt}
--no parenthesis--, so that advanced users could use the code as a template.
}
\details{
The pipeline combines \code{\link{find_sky_dns}},
\code{\link{fit_cone_shaped_model}}, \code{\link{choose_std_cie_sky_model}},
\code{\link{fit_trend_surface}}, and \code{\link{thr_image}}. The conceptual
design is the same that for \code{\link{ootb_mblt}}, but the working
binarized image produced by \code{\link{fit_cone_shaped_model}} is refined by
\code{\link{choose_std_cie_sky_model}}, and the standard CIE sky model is the
filling source for \code{\link{fit_trend_surface}}, instead of the cone
shaped model.

Also, the sky actually used to obtain the binarized image is from the
combination of the standard CIE sky model and the trend surface, instead of
the cone shaped model and the trend surface.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
bin <- ootb_cie_mblt(blue, z, a)$bin
plot(bin)
}
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
Other mblt functions: 
\code{\link{find_sky_dns}()},
\code{\link{fit_cone_shaped_model}()},
\code{\link{fit_trend_surface}()},
\code{\link{ootb_mblt}()},
\code{\link{thr_image}()}
}
\concept{mblt functions}
