% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ootb_obia.R
\name{ootb_obia}
\alias{ootb_obia}
\title{Out-of-the-box object-based image analysis of hemispherical photographs}
\usage{
ootb_obia(caim, z, a, m, sky_blue)
}
\arguments{
\item{caim}{\linkS4class{SpatRaster}. The return of a call to
\code{\link{read_caim}}.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link{zenith_image}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link{azimuth_image}}.}

\item{m}{\linkS4class{SpatRaster}. A mask. Usually, built with
\code{\link{mask_hs}}.}

\item{sky_blue}{\linkS4class{color}. Is the \code{target_color} argument to
be passed to \code{\link{membership_to_color}}.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and
  \code{1}.
}
\description{
Out-of-the-box version of methods first presented in
\insertCite{Diaz2015;textual}{rcaiman}.
}
\details{
This function is a hard-coded version of a pipeline that combines these main
functions \code{\link{enhance_caim}}, \code{\link{polar_qtree}},
\code{\link{obia}}. The code can be easily inspected by calling
\code{ootb_obia} --no parenthesis. Advanced users can use that code as a
template.

If you use this function in your research, please cite
\insertCite{Diaz2015;textual}{rcaiman} in addition to this package.
}
\examples{
\dontrun{
caim <- read_caim() \%>\% normalize()
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
sky_blue_sample <- crop(caim, ext(610,643,760,806))
sky_blue <- apply(sky_blue_sample[], 2, median) \%>\%
  as.numeric() \%>\%
  matrix(., ncol = 3) \%>\%
  sRGB()
bin <- ootb_obia(caim, z, a, !is.na(z), sky_blue)
plot(bin)
}
}
\seealso{
Other Binarization Functions: 
\code{\link{apply_thr}()},
\code{\link{find_sky_pixels_nonnull}()},
\code{\link{find_sky_pixels}()},
\code{\link{obia}()},
\code{\link{ootb_mblt}()},
\code{\link{regional_thresholding}()},
\code{\link{thr_image}()}
}
\concept{Binarization Functions}
