% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_sun_zenith_azimuth.R
\name{optim_sun_zenith_azimuth}
\alias{optim_sun_zenith_azimuth}
\title{Optimize sun angular coordinates}
\usage{
optim_sun_zenith_azimuth(sun_zenith_azimuth, rr, sky_coef, method = "BFGS")
}
\arguments{
\item{sun_zenith_azimuth}{Numeric vector of lenght two. See
\code{\link[=extract_sun_zenith_azimuth]{extract_sun_zenith_azimuth()}}. See also \code{\link[=row_col_from_zenith_azimuth]{row_col_from_zenith_azimuth()}} in
case you want to provide values based on date and time of acquisition and
the \code{suncalc} package.}

\item{rr}{An object of class \emph{list}. The output of \code{\link[=extract_rel_radiance]{extract_rel_radiance()}}
or an object with same structure and names.}

\item{sky_coef}{Numeric vector of length five. Parameters of the sky model.}

\item{method}{The method to be used. See \sQuote{Details}.  Can be abbreviated.}
}
\value{
Numeric vector of length two, where the first element is the
solar zenith angle and the second is the solar azimuth angle, both
expressed in degrees.
}
\description{
Optimize sun angular coordinates
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

# See fit_cie_sky_model() for details on below file
path <- system.file("external/sky_points.csv",
                    package = "rcaiman")
sky_points <- read.csv(path)
sky_points <- sky_points[c("Y", "X")]
colnames(sky_points) <- c("row", "col")
head(sky_points)
plot(caim$Blue)
points(sky_points$col, nrow(caim) - sky_points$row, col = 2, pch = 10)

plot(caim$Blue)
points(sky_points$col, nrow(caim) - sky_points$row, col = 2, pch = 10)

# x11()
# plot(caim$Blue)
# sun_zenith_azimuth <- click(c(z, a), 1) \%>\% unname()
sun_zenith_azimuth <- c(49.5, 27.42481) #taken with above lines then hardcoded

sun_row_col <- row_col_from_zenith_azimuth(z, a,
                                           sun_zenith_azimuth[1],
                                           sun_zenith_azimuth[2])
points(sun_row_col[2], nrow(caim) - sun_row_col[1], col = 3, pch = 1)

rr <- extract_rel_radiance(caim$Blue, z, a, sky_points)

set.seed(7)
model <- fit_cie_sky_model(rr, sun_zenith_azimuth,
                           general_sky_type = "Clear",
                           twilight = 90,
                           method = "BFGS")

sun_zenith_azimuth <- optim_sun_zenith_azimuth(c(49.5, 27.42481), rr,
                                               model$coef, method = "CG")
sun_zenith_azimuth
sun_row_col <- row_col_from_zenith_azimuth(z, a,
                                           sun_zenith_azimuth[1],
                                           sun_zenith_azimuth[2])
points(sun_row_col[2], nrow(caim) - sun_row_col[1], col = 4, pch = 1)
}
}
