% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_sampling_uniformity.R
\name{optimize_sampling_uniformity}
\alias{optimize_sampling_uniformity}
\title{Optimize the minimum distance between sky sampling points}
\usage{
optimize_sampling_uniformity(sky_points, r, z, a, equalarea_seg, min_dist_seq)
}
\arguments{
\item{sky_points}{\code{data.frame} with columns \code{row} and \code{col} (raster
coordinates).}

\item{r}{\link[terra:SpatRaster-class]{terra::SpatRaster}. Image from which \code{sky_points} were
sampled (or any raster with identical dimensions).}

\item{z}{\link[terra:SpatRaster-class]{terra::SpatRaster} generated with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\link[terra:SpatRaster-class]{terra::SpatRaster} generated with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{equalarea_seg}{\link[terra:SpatRaster-class]{terra::SpatRaster} generated with \code{\link[=equalarea_segmentation]{equalarea_segmentation()}}.}

\item{min_dist_seq}{numeric vector. Candidate minimum distances (deg).}
}
\value{
List with:
\describe{
\item{\code{sky_points}}{Thinned point set using the optimal distance.}
\item{\code{min_dist}}{Value in \code{min_dist_seq} that minimized KL divergence.}
\item{\code{kl_values}}{KL divergence for each candidate distance.}
}
}
\description{
Evaluate a sequence of candidate minimum distances and select the one that
minimizes the Kullback–Leibler (KL) divergence from a uniform distribution on
an equal-area hemispherical segmentation.
}
\details{
For each value in \code{min_dist_seq}, the input point pattern (\code{sky_points}) is
thinned using \code{\link[=rem_nearby_points]{rem_nearby_points()}}. The resulting points are assigned to the
cells of \code{equalarea_seg}, and their empirical distribution is compared with
the uniform reference through KL divergence, as computed by
\code{\link[=assess_sampling_uniformity]{assess_sampling_uniformity()}}. As a results, it improves spatial regularity
by selecting an optimal minimum distance.
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)
m <- !is.na(z)
r <- caim$Blue

bin <- binarize_by_region(r, ring_segmentation(z, 15), "thr_isodata") &
  select_sky_region(z, 0, 88)

seg <- skygrid_segmentation(z, a, 10)
sky_points <- sample_sky_points(r, bin, seg,
                                dist_to_black = 3)
plot(bin)
points(sky_points$col, nrow(caim) - sky_points$row, col = 2, pch = 10)

l <- optimize_sampling_uniformity(sky_points, NULL, z, a, equalarea_segmentation(z, a, 100), 0:15)
plot(l$kl_values)
l$min_dist

plot(bin)
points(l$sky_points$col, nrow(caim) - l$sky_points$row, col = 2, pch = 10)
}
}
