% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_of_clipped_highlights.R
\name{percentage_of_clipped_highlights}
\alias{percentage_of_clipped_highlights}
\title{Percentage of clipped highlights}
\usage{
percentage_of_clipped_highlights(r, m)
}
\arguments{
\item{r}{Single-layer object from the \linkS4class{SpatRaster}.}

\item{m}{\linkS4class{SpatRaster}. A mask. For hemispherical photographs,
check \code{\link[=select_sky_vault_region]{select_sky_vault_region()}}.}
}
\value{
Numeric vector of lenght one.
}
\description{
Wrapper function for \code{\link[terra:freq]{terra::freq()}}
}
\examples{
r <- read_caim()$Blue
z <- zenith_image(ncol(r), lens())
m <- !is.na(z)
percentage_of_clipped_highlights(r, m)
r <- normalize_minmax(r, 0, 1000, TRUE)
percentage_of_clipped_highlights(r, m)
}
