% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtree.R
\name{qtree}
\alias{qtree}
\title{Do quad-tree segmentation}
\usage{
qtree(r, scale_parameter = 0.2)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}.}

\item{scale_parameter}{Numeric vector of length one. Quad-tree is a top-down
method. This parameter controls the stopping condition. Therefore, it
allows controlling the size of the resulting segments. Ultimately, segments
sizes will depend on both this parameter and the heterogeneity of \code{r}.}
}
\value{
A single layer image of the class \linkS4class{SpatRaster} with integer
values.
}
\description{
The quad-tree segmentation algorithm is a top-down process that makes
recursive divisions in four equal parts until a condition is satisfied and
stops locally. This is the usual implementation of the quad-tree algorithm,
so it produces squared segments of different sizes. This particular
implementation allows up to five sizes.
}
\details{
The algorithm starts splitting the entire image into large squared segments.
Depending on the aspect ratio, starting grids will going from \eqn{4 \times
4} to \eqn{1 \times 4} or \eqn{4 \times 1}. Then, it splits each segment into
four sub-segments and calculates the standard deviation of the pixels from
\code{r} delimited by each of those sub-segments and segment. The splitting
process stops locally if \emph{delta}, the sum of the standard deviation of the
sub-segments minus the standard deviation of the parent segment, is less or
equal than the \code{scale_parameter}. If \code{r} has more than one layer,
\emph{delta} is calculated separately and \emph{delta} mean is used to
evaluate the stopping condition.
}
\examples{
\dontrun{
caim <- read_caim() \%>\% normalize_minmax()
seg <- qtree(caim, scale_parameter = 0.05)
plot(caim$Blue)
plot(extract_feature(caim$Blue, seg))
plot(extract_feature(seg, seg, length))
}
}
