% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcaiman-package.R
\docType{package}
\name{rcaiman-package}
\alias{rcaiman}
\alias{rcaiman-package}
\title{rcaiman: CAnopy IMage ANalysis}
\description{
Tools for preprocessing and analysis of canopy photographs.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/GastonMauroDiaz/rcaiman/issues}
}

}
\author{
\strong{Maintainer}: Gastón Mauro Díaz \email{gastonmaurodiaz@gmail.com} (\href{https://orcid.org/0000-0002-0362-8616}{ORCID})

}
\keyword{internal}
