% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_caim.R
\name{read_caim}
\alias{read_caim}
\title{Read a canopy image from a file}
\usage{
read_caim(path = NULL, zenith_colrow = NULL, horizon_radius = NULL)
}
\arguments{
\item{path}{character vector of length one.  Path to an image file, including
extension. If \code{NULL}, an example image is returned.}

\item{zenith_colrow}{numeric vector of length 2. Column and row coordinates
of the optical center (zenith) in pixel units.}

\item{horizon_radius}{integer-like numeric vector of length one. Radius of the
hemispherical image in pixel units, measured from the optical center
(\code{zenith_colrow}) to the image location corresponding to a zenith angle of
90 deg (i.e. the horizon). It must be an even integer (see
\link{zenith_image}).}
}
\value{
Numeric \link[terra:SpatRaster-class]{terra::SpatRaster}, typically with layers named
\code{"Red"}, \code{"Green"}, and \code{"Blue"}. If the file extension does not correspond
to typical JPEG or TIFF files, names will be inferred and a
warning will be issued.
}
\description{
Reads a born-digital image (typically RGB-JPEG or RGB-TIFF) using
\code{\link[terra:rast]{terra::rast()}} and returns a \link[terra:SpatRaster-class]{terra::SpatRaster} object.
}
\details{
Internally, this is a wrapper around \code{\link[terra:rast]{terra::rast()}}, so support for image
formats depends on the capabilities of the \code{terra} package.

This function is intended for importing color hemispherical photographs, such
as those obtained with digital cameras equipped with fisheye lenses. For raw
image files (e.g., NEF, CR2), see \code{\link[=read_caim_raw]{read_caim_raw()}}. A typical workflow makes
use of \code{\link[=read_caim_raw]{read_caim_raw()}}, \code{\link[=write_caim]{write_caim()}}, and \code{\link[=read_caim]{read_caim()}}

The example image was created from a raw photograph taken with a Nikon Coolpix
5700 and a FC-E9 auxiliary lens, processed with the following code:

\if{html}{\out{<div class="sourceCode">}}\preformatted{zenith_colrow <- c(1290, 988)/2
diameter <- 756
z <- zenith_image(diameter, lens("Nikon_FCE9"))
a <- azimuth_image(z)
m <- !is.na(z)
caim <- read_caim_raw("DSCN4606.NEF")
caim <- crop_caim(caim, zenith_colrow - diameter/2, diameter, diameter)
caim <- correct_vignetting(caim, z, c(0.0638, -0.101))
caim <- c(mean(caim$Y, caim$M), caim$G, caim$C)
caim <- fisheye_to_equidistant(caim, z, a, m, radius = 300, k = 1)
write_caim(caim, "example.tif", 16)
}\if{html}{\out{</div>}}
}
\note{
\code{rcaiman} uses terra without geographic semantics: rasters are kept with
unit resolution (cell size = 1) and a standardized extent
\code{ext(0, ncol, 0, nrow)} with CRS EPSG:7589 (\code{\link[=set_rcaiman_geometry]{set_rcaiman_geometry()}}).
}
\examples{
path <- system.file("external/DSCN4500.JPG", package = "rcaiman")
zenith_colrow <- c(1286, 986)
horizon_radius <- 742
caim <- read_caim(path, zenith_colrow, horizon_radius)
plot(caim$Blue)
}
\seealso{
\code{\link[=write_caim]{write_caim()}}
}
