% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ootb_sky_model.R
\name{read_ootb_sky_model}
\alias{read_ootb_sky_model}
\title{Read files writen by \code{\link[=write_ootb_sky_model]{write_ootb_sky_model()}}}
\usage{
read_ootb_sky_model(name, z, a)
}
\arguments{
\item{name}{Character vector of length one. File name without extension. If
needed, a file path can be added to the names. For example:
\code{"C:/Users/Doe/Documents/DSCN4500"}.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}
}
\value{
An object of the class \emph{list} similar to the output of
\code{\link[=ootb_fit_cie_sky_model]{ootb_fit_cie_sky_model()}}.
}
\description{
Read files writen by \code{\link[=write_ootb_sky_model]{write_ootb_sky_model()}}
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

path <- system.file("external/ootb_sky.txt", package = "rcaiman")
ootb_sky <- read_ootb_sky_model(gsub(".txt", "", path), z, a)
}
}
