% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{row_col_from_zenith_azimuth}
\alias{row_col_from_zenith_azimuth}
\title{Row and col numbers from zenith and azimuth angles}
\usage{
row_col_from_zenith_azimuth(r, za, lens_coef)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{za}{Numeric vector of length two. Zenith and azimuth angles in degrees.}
}
\description{
Row and col numbers from zenith and azimuth angles
}
\examples{
z <- zenith_image(1000, lens_coef = lens())
row_col <- row_col_from_zenith_azimuth(z, c(45, 270), lens())
# the opposite calculation does not require a function
a <- azimuth_image(z)
z[row_col[1], row_col[2]]
a[row_col[1], row_col[2]]
}
\seealso{
Other hps functions: 
\code{\link{extract_sky_marks}()}
}
\concept{hps functions}
