% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_circumsolar_region.R
\name{select_circumsolar_region}
\alias{select_circumsolar_region}
\title{Select circumsolar region}
\usage{
select_circumsolar_region(z, a, sun_zenith_azimuth, chi_max_sun)
}
\arguments{
\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{sun_zenith_azimuth}{Numeric vector of lenght two. See
\code{\link[=extract_sun_zenith_azimuth]{extract_sun_zenith_azimuth()}}.}

\item{chi_max_sun}{Numeric vector of length one. Specifies the maximum
expected size of the circumsolar region in degrees.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and
\code{1}.
}
\description{
Select circumsolar region
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)
m <- !is.na(z)
r <- caim$Blue

sun_zenith_azimuth <- extract_sun_zenith_azimuth(r, z, a, m, NULL)
m_sun <- select_circumsolar_region(z, a, sun_zenith_azimuth, 20)
plot(m_sun)
}
}
