% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sky_vault_region.R
\name{select_sky_vault_region}
\alias{select_sky_vault_region}
\title{Select a portion of the sky vault}
\usage{
select_sky_vault_region(r, from, to)
}
\arguments{
\item{r}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}} or
\code{\link[=azimuth_image]{azimuth_image()}}.}

\item{from, to}{angle in degrees, inclusive limits.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and
\code{1}.
}
\description{
Given a zenith or azimuth image and angle restrictions, this function
produces select a region of the sky vault and produce a binary raster mask.
}
\examples{
\dontrun{
z <- zenith_image(1000, lens())
a <- azimuth_image(z)
m1 <- select_sky_vault_region(z, 20, 70)
plot(m1)
m2 <- select_sky_vault_region(a, 330,360)
plot(m2)
plot(m1 & m2)
plot(m1 | m2)

# 15 degrees at each side of 0
m1 <- select_sky_vault_region(a, 0, 15)
m2 <- select_sky_vault_region(a, 345, 360)
plot(m1 | m2)

# better use this
plot(!is.na(z))
# instead of this
plot(select_sky_vault_region(z, 0, 90))
}
}
\seealso{
\code{\link[=masking]{masking()}}
}
