% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sky_grid_points.R
\name{sky_grid_points}
\alias{sky_grid_points}
\title{Cell central points of a sky grid}
\usage{
sky_grid_points(z, a, angle_width, lens_coef = lens())
}
\arguments{
\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{angle_width}{Numeric vector of length one. It should be \verb{30, 15, 10, 7.5, 6, 5, 3.75, 3, 2.5, 1.875, 1} or \code{0.5} degrees. This constrain is
rooted in the requirement of a value able to divide both the \code{0} to \code{360}
and \code{0} to \code{90} ranges into a whole number of segments.}

\item{lens_coef}{Numeric vector. Polynomial coefficients of the lens
projection function. See \code{\link[=calibrate_lens]{calibrate_lens()}}.}
}
\value{
data.frame
}
\description{
Cell central points of a sky grid
}
\examples{
\dontrun{
}
}
