% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblt_subpixel.R
\name{subpixel_to_bin}
\alias{subpixel_to_bin}
\title{Subpixel to binary classification}
\usage{
subpixel_to_bin(subpixel, segmentation)
}
\arguments{
\item{subpixel}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{mblt_subpixel}}.}

\item{segmentation}{\linkS4class{RasterLayer}.This should match with the type
of sky grid that will be used in the software that goes next in the
processing pipeline.}
}
\description{
Transform a subpixel classification from \code{mblt_subpixel} into a
binarized image.
}
\details{
This function is useful to export the result of \code{\link{mblt_subpixel}}
to a software that does not handle subpixel classification, such as Hemisfer
and CIMES.
}
