% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblt_1.R
\name{thr_fun}
\alias{thr_fun}
\title{Threshold function}
\usage{
thr_fun(dn, w = 0.5, type = "Generic", intercept = NULL, slope = NULL)
}
\arguments{
\item{dn}{Numeric vector. Digital number of the background. The background
should be lighter than the objects. In canopy photography, the typical
background is the sky.}

\item{w}{Numeric vector of length one. Weight. See reference.}

\item{type}{Character vector of length one. Default is "Generic". Currently,
the only available calibrated values are from
\insertCite{Diaz2018}{rcaiman}. Use "Nikon_Coolpix_5700" to use them.}

\item{intercept, slope}{Numeric vector of length one. Default is NULL. You can
provide your own calibrated values. See \insertCite{Diaz2018}{rcaiman} for
details.}
}
\description{
Threshold value in function of the background digital number.
}
\examples{
thr_fun(125)
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
