% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ootb_sky_model.R
\name{write_ootb_sky_model}
\alias{write_ootb_sky_model}
\title{Write out-of-the-box CIE sky model fitting}
\usage{
write_ootb_sky_model(ootb_sky, name)
}
\arguments{
\item{ootb_sky}{An object of the class \code{list} that is the result of calling
\code{\link[=ootb_fit_cie_sky_model]{ootb_fit_cie_sky_model()}}.}

\item{name}{Character vector of length one. File name without extension. If
needed, a file path can be added to the names. For example:
\code{"C:/Users/Doe/Documents/DSCN4500"}.}
}
\value{
No return value. Called for side effects.
}
\description{
Write out-of-the-box CIE sky model fitting
}
