% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Write report}
\usage{
write_report(ootb_sky, file)
}
\arguments{
\item{ootb_sky}{An object of the class \code{list} that is the result of calling
\code{\link[=ootb_sky_reconstruction]{ootb_sky_reconstruction()}}.}

\item{file}{a writable \link[base]{connection} or a character string naming the
    file to write to, or \code{NULL} to stop sink-ing.}
}
\value{
No return value. Called for side effects.
}
\description{
Write report using \code{\link[base:cat]{base::cat()}} and \code{\link[base:sink]{base::sink()}}.
}
\seealso{
Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{calc_sngd}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{expand_sky_points}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rl}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_general_sky_type}()},
\code{\link{find_sky_pixels}()},
\code{\link{masking}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{sor_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()}
}
\concept{Tool Functions}
