% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{write_sky_marks}
\alias{write_sky_marks}
\title{Write sky marks}
\usage{
write_sky_marks(x, path_to_HSP_project, img_name)
}
\arguments{
\item{x}{Object from the class data.frame. The result of a calling to
\code{\link{extract_sky_marks}}.}

\item{path_to_HSP_project}{Character vector of length one. Path to the HSP
project folder.}

\item{img_name}{Character vector of length one. See details.}
}
\value{
None. A file will be written in the HSP project folder.
}
\description{
Create a special file to interface with HSP software.
}
\details{
This function is part of a workflow that connects the mblt algorithm with the
HSP software package \insertCite{Lang2013}{rcaiman}.

This function was designed to be called after \code{\link{extract_sky_marks}}
in a workflow that connects the mblt algorithm with the HSP software package.
I such a workflow, the \code{\link{extract_sky_marks}} will use an image
pre-processed by the HSP software as its \code{r} argument. The
\code{img_name} argument of \code{write_sky_marks()} should be the name of
the file associated to the aforementioned \code{r} argument.
}
\references{
\insertRef{Lang2013}{rcaiman}
}
\seealso{
Other hsp functions: 
\code{\link{extract_sun_mark}()},
\code{\link{write_sun_mark}()}
}
\concept{hsp functions}
