% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{write_sky_points}
\alias{write_sky_points}
\title{Write sky points}
\usage{
write_sky_points(x, path_to_HSP_project, img_name)
}
\arguments{
\item{x}{An object of the class \emph{data.frame}. The result of a calling to
\code{\link{extract_sky_points}}.}

\item{path_to_HSP_project}{Character vector of length one. Path to the HSP
project folder.}

\item{img_name}{Character vector of length one. See details.}
}
\value{
None. A file will be written in the HSP project folder.
}
\description{
Create a special file to interface with the HSP software.
}
\details{
This function is part of a workflow that connects this package with the HSP
software package \insertCite{Lang2013}{rcaiman}.

This function was designed to be called after
\code{\link{extract_sky_points}}, and is part of a workflow that connects the
MBLT algorithm with the HSP software package. In such a workflow, the
\code{r} argument provided to \code{\link{extract_sky_points}} should be an
image pre-processed by the HSP software. Those images are stored as PGM files
by HSP in the subfolder "manipulate" of the project folder (which will be in
turn a subfolder of the "project\strong{s}" folder). Those PGM files can be
read with \code{\link[terra]{rast}}. For instance: \code{r <-
rast("C:/Users/johndoe/Documents/HSP/Projects/my_prj/manipulate/img01.pgm")}.
Then, they will be ready for use as input after running
\code{normalize(r)}.

The \code{img_name} argument of \code{write_sky_points()} should be the name
of the file associated to the aforementioned \code{r} argument.
}
\references{
\insertAllCited{}
}
\seealso{
Other HSP Functions: 
\code{\link{extract_sun_coord}()},
\code{\link{read_manual_input}()},
\code{\link{read_opt_sky_coef}()},
\code{\link{row_col_from_zenith_azimuth}()},
\code{\link{write_sun_coord}()},
\code{\link{zenith_azimuth_from_row_col}()}
}
\concept{HSP Functions}
