% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{write_sun_coord}
\alias{write_sun_coord}
\title{Write sun coordinates}
\usage{
write_sun_coord(x, path_to_HSP_project, img_name)
}
\arguments{
\item{x}{Numeric vector of length two. Raster coordinates of the solar disk
that can be obtained by calling to \code{\link{extract_sun_coord}}.
\strong{TIP}: if the output of \code{extrac_sun_coord()} is
\code{sun_coord}, then you should provide to \code{write_sun_coord()} this:
\code{sun_coord$row_col}. See also
\code{\link{row_col_from_zenith_azimuth}}.}

\item{path_to_HSP_project}{Character vector of length one. Path to the HSP
project folder.}

\item{img_name}{Character vector of length one. See details.}
}
\value{
None. A file will be written in the HSP project folder.
}
\description{
Create a special file to interface with the HSP software.
}
\details{
Refer to the Details section of this function:
\code{\link{write_sky_points}}.
}
\seealso{
Other HSP Functions: 
\code{\link{extract_sun_coord}()},
\code{\link{read_manual_input}()},
\code{\link{read_opt_sky_coef}()},
\code{\link{row_col_from_zenith_azimuth}()},
\code{\link{write_sky_points}()},
\code{\link{zenith_azimuth_from_row_col}()}
}
\concept{HSP Functions}
