% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{write_sun_mark}
\alias{write_sun_mark}
\title{Write sun mark}
\usage{
write_sun_mark(x, path_to_HSP_project, img_name)
}
\arguments{
\item{x}{Numeric vector of lenght two. Raster coordinates of the solar disk
that can be obtained by calling to \code{\link{extract_sun_mark}}. *TIP*:
if the output of \code{extrac_sun_mark()} is \code{x}, then you should
provide to \code{write_sun_mark()} this: \code{x$row_col}. See also
\code{\link{row_col_from_zenith_azimuth}}.}

\item{path_to_HSP_project}{Character vector of length one. Path to the HSP
project folder.}

\item{img_name}{Character vector of length one. See details.}
}
\value{
None. A file will be written in the HSP project folder.
}
\description{
Create a special file to interface with HSP software.
}
\details{
Please, see the Details section of this function:
\code{\link{write_sky_marks}}.
}
\seealso{
Other hsp functions: 
\code{\link{extract_sun_mark}()},
\code{\link{write_sky_marks}()}
}
\concept{hsp functions}
