% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_thr.R
\name{apply_thr}
\alias{apply_thr}
\title{Apply threshold}
\usage{
apply_thr(r, thr)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. Image to binarize.}

\item{thr}{Numeric vector of length one or \linkS4class{SpatRaster}.
Threshold.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and
  \code{1}.
}
\description{
Global or local thresholding of images.
}
\details{
It is a wrapper function around the operator \code{>} from the ‘terra’
package. If a single threshold value is provided as \code{thr} argument, it
is applied to every pixel of the object \code{r}. If instead a
\linkS4class{SpatRaster} is provided as \code{thr} argument, then a
particular threshold is applied to each particular pixel.
}
\examples{
r <- read_caim()
apply_thr(r$Blue, 120)
\dontrun{
# This function is useful in combination with the ‘autothresholdr’
# package. For example:
require(autothresholdr)
thr <- auto_thresh(r$Blue[], "IsoData")[1]
bin <- apply_thr(r$Blue, thr)
plot(bin)
}
}
\seealso{
Other Binarization functions: 
\code{\link{find_sky_pixels_nonnull_criteria}()},
\code{\link{find_sky_pixels}()},
\code{\link{ootb_mblt}()}
}
\concept{Binarization functions}
