% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_co.R
\name{calc_co}
\alias{calc_co}
\title{Calculate canopy openness}
\usage{
calc_co(bin, m, z, a, angle_width)
}
\arguments{
\item{bin}{\linkS4class{SpatRaster}. Binarized hemispherical canopy image.}

\item{m}{\linkS4class{SpatRaster}. Check \code{\link{mask_hs}}.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link{zenith_image}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link{azimuth_image}}.}

\item{angle_width}{Numeric vector of length one. It should be \code{30, 15,
10, 7.5, 6, 5, 3.75, 3, 2.5, 1.875, 1} or \code{0.5} degrees. This
constrain is rooted in the requirement of a value able to divide both the
\code{0} to \code{360} and \code{0} to \code{90} ranges into a whole number
of segments.}
}
\value{
Numeric vector of length one.
}
\description{
Calculate canopy openness
}
\details{
Canopy openness calculated as in the equation from
\insertCite{Gonsamo2011;textual}{rcaiman}:

\eqn{CO = \sum_{i = 1}^{N} GF(\phi_i, \theta_i) \cdot ((cos(\theta_1) -
cos(\theta_2))/n)}, where \eqn{GF(\phi_i, \theta_i)} is the gap fraction of
the cell \eqn{i}, \eqn{\theta_1} and \eqn{\theta_2} are the smallest and
largest zenith angle of the cell \eqn{i}, and \eqn{n} is the number of cells
on the ring delimited by \eqn{\theta_1} and \eqn{\theta_2}.
}
\examples{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
m <- !is.na(z)
bin <- apply_thr(caim$Blue, thr_isodata(caim$Blue[m]))
plot(bin)
calc_co(bin, m, z, a, 10)

}
\references{
\insertAllCited{}
}
\seealso{
Other Tool Functions: 
\code{\link{colorfulness}()},
\code{\link{defuzzify}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rl}()},
\code{\link{extract_sky_points}()},
\code{\link{masking}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()}
}
\concept{Tool Functions}
