% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_functions.R
\name{calc_diameter}
\alias{calc_diameter}
\title{Calculate diameter}
\usage{
calc_diameter(lens_coef, radius_px, angle)
}
\arguments{
\item{lens_coef}{Numeric vector. Polynomial coefficients
of the lens projection function.}

\item{radius_px}{Numeric vector. Distance in pixels from the zenith.}

\item{angle}{Numeric vector. Zenith angle in degrees.}
}
\description{
Calculate the diameter of a 180º fisheye image.
}
\details{
This function is useful to handle devices with field of view different than
180 degrees. Given a lens projection function and data points consisting of
radii (pixels) and their correspondent zenith angle, it returns the radius of
the horizon (i.e., the radius for the zenith angle equal to 90 degrees).

It is particularly useful when working with non-circular hemispherical
photography. It will help to find the diameter that a circular image would
have if the equipment would depict the whole hemisphere.
}
\examples{
# Nikon D50 and Fisheye Nikkor 10.5 mm lens
calc_diameter(lens("Nikkor_10.5_mm"), 1202, 53)
}
