% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_oor_index.R
\name{calc_oor_index}
\alias{calc_oor_index}
\title{Calc out-of-range index}
\usage{
calc_oor_index(r, sky)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. The blue channel of a canopy photograph.}

\item{sky}{\linkS4class{SpatRaster}. The blue radiance of the unobscured sky}
}
\value{
Numeric vector of length one.
}
\description{
The \strong{out-out-of-range} (OOR) index is calculated as foollow:
}
\details{
\eqn{\sum_{i = 1}^{N}(r_i/sky_i)^2},

where \eqn{r} is a canopy image with radiance data, \eqn{sky} is an
image with unobstructed sky radiance data, \eqn{i} is the index that
represents the position of a given pixel on the raster grid, and \eqn{N} is
the total number of pixels that satisfy: \eqn{r_i/sky_i<0} or
\eqn{r_i/sky_i>1}.
}
\examples{
\dontrun{
caim <- read_caim()
r <- caim$Blue
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

path <- system.file("external/ootb_sky.txt", package = "rcaiman")
ootb_sky <- read_ootb_sky_model(gsub(".txt", "", path), z, a)

sky <- ootb_interpolate_and_merge(r, z, a, ootb_sky$sky_points, ootb_sky)

calc_oor_index(r, sky$sky)
}
}
