% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_oor_index.R
\name{calc_oor_index}
\alias{calc_oor_index}
\title{Calc out-of-range index}
\usage{
calc_oor_index(r, sky)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically, the
blue channel extracted from a canopy photograph. Please see \code{\link[=read_caim]{read_caim()}}
and \code{\link[=normalize]{normalize()}}.}

\item{sky}{\linkS4class{SpatRaster}. The relative radiance of the unobscured sky}
}
\value{
Numeric vector of length one.
}
\description{
The \strong{out-out-of-range} (OOR) index is calculated as foollow:
}
\details{
\eqn{\sum_{i = 1}^{N}(r_i/sky_i)^2},

where \eqn{r} is a canopy image with relative radiance data, \eqn{sky} is an
image with unobstructed sky relative radiance data, \eqn{i} is the index that
represents the position of a given pixel on the raster grid, and \eqn{N} is
the total number of pixels that satisfy: \eqn{r_i/sky_i<0} or
\eqn{r_i/sky_i>1}.
}
\examples{
\dontrun{
path <- system.file("external/DSCN4500.JPG", package = "rcaiman")
caim <- read_caim(path, c(1250, 1020) - 745, 745 * 2, 745 * 2)
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
r <- gbc(caim$Blue)
r <- correct_vignetting(r, z, c(0.0638, -0.101)) \%>\% normalize()

bin <- regional_thresholding(r, rings_segmentation(z, 30), "thr_isodata")
bin <- bin & mask_hs(z, 0, 80)
sky_points <- extract_sky_points(r, bin, sky_grid_segmentation(z, a, 3))
sky_points <- extract_rl(r, z, a, sky_points, no_of_points = NULL)

model <- fit_coneshaped_model(sky_points$sky_points)
summary(model$model)
sky_cs <- model$fun(z, a)
plot(r/sky_cs)
calc_oor_index(r, sky_cs)
}
}
\seealso{
Other Tool Functions: 
\code{\link{calc_sngd}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{expand_sky_points}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rl}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_general_sky_type}()},
\code{\link{find_sky_pixels}()},
\code{\link{masking}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{read_ootb_sky_model}()},
\code{\link{sor_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
