% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative_radius.R
\name{calc_relative_radius}
\alias{calc_relative_radius}
\title{Calculate relative radius}
\usage{
calc_relative_radius(angle, lens_coef)
}
\arguments{
\item{angle}{Numeric vector. Zenith angles in degrees.}

\item{lens_coef}{Numeric vector. Polynomial coefficients of the lens
projection function. See \code{\link[=calibrate_lens]{calibrate_lens()}}.}
}
\description{
Calculate the relative radius given a zenith angle and lens function. This is
known as the projection function.
}
