% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessboard.R
\name{chessboard}
\alias{chessboard}
\title{Do chessboard segmentation}
\usage{
chessboard(r, size)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}.}

\item{size}{Numeric vector of length one. Size of the square segments.}
}
\value{
A single layer image of the class \linkS4class{SpatRaster} with integer
values.
}
\description{
Do chessboard segmentation
}
\examples{
caim <- read_caim()
seg <- chessboard(caim, 20)
plot(caim$Blue)
plot(extract_feature(caim$Blue, seg))
}
