% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cie_sky_model.R
\name{choose_std_cie_sky_model}
\alias{choose_std_cie_sky_model}
\title{Choose a standard CIE sky model}
\usage{
choose_std_cie_sky_model(r, z, a, bin)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{RasterLayer}. The result of binarizing the \code{r}
argument.}
}
\description{
Choose one from the 15th standard CIE sky models based on sky DN values.
}
\details{
This function assume the same than \code{\link{find_sky_dns}}, so please
refer to that function help, first paragraph under Details headline.

To kwow more about the CIE sky models, please refer to
\insertCite{Li2016;textual}{rcaiman}.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
sky <- choose_std_cie_sky_model(blue, z, a)$relative_luminance
plot(sky)
}
}
\references{
\insertRef{Li2016}{rcaiman}
}
\seealso{
Other cie sky model functions: 
\code{\link{autofit_cie_sky_model}()},
\code{\link{cie_sky_model_raster}()},
\code{\link{fit_cie_sky_model}()}
}
\concept{cie sky model functions}
