% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_vignetting.R
\name{correct_vignetting}
\alias{correct_vignetting}
\title{Correct vignetting effect}
\usage{
correct_vignetting(r, z, lens_coef_v)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A fish-eye image.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{lens_coef_v}{Numeric vector. Coefficients of a vignetting
function (\eqn{f_v}) of the type \eqn{f_v = 1 + a \cdot \theta +
  b \cdot \theta^2 + ... + m \cdot \theta^n}, where \eqn{\theta} is the
zenith angle, \eqn{a, b, c} and \eqn{m} are the coefficients. The maximum
polynomial degree supported is sixth. See \code{\link[=extract_radiometry]{extract_radiometry()}} for
additional details.}
}
\value{
The argument \code{r} but with corrected values.
}
\description{
Correct vignetting effect
}
\examples{
\dontrun{
path <- system.file("external/APC_0581.jpg", package = "rcaiman")
caim <- read_caim(path)
z <- zenith_image(2132/2,  c(0.7836, 0.1512, -0.1558))
a <- azimuth_image(z)
zenith_colrow <- c(1063, 771)/2

caim <- expand_noncircular(caim, z, zenith_colrow)
m <- !is.na(caim$Red) & !is.na(z)
caim[!m] <- 0

bin <- apply_thr(caim$Blue, thr_isodata(caim$Blue[m]))

display_caim(caim$Blue, bin)

caim <- gbc(caim, 2.2)
caim <- correct_vignetting(caim, z, c(-0.0546, -0.561, 0.22)) \%>\%
                                                     normalize_minmax()
# The lens_coef_v are here coef 10.1016/j.agrformet.2024.110020
}
}
