% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defuzzify.R
\name{defuzzify}
\alias{defuzzify}
\title{Defuzzify a fuzzy classification}
\usage{
defuzzify(mem, segmentation)
}
\arguments{
\item{mem}{A \linkS4class{SpatRaster} object representing the degree of membership
in a fuzzy classification.}

\item{segmentation}{An object of the class \linkS4class{SpatRaster} such as an
object returned by \code{\link[=sky_grid_segmentation]{sky_grid_segmentation()}}.}
}
\value{
An object of the class \linkS4class{SpatRaster} containing binary
information.
}
\description{
This function translates degree of membership into Boolean logic using a
regional approach. The output ensures that the fuzzy and Boolean versions
remain consistent at the specified level of aggregation (controlled by the
argument \code{segmentation}). This method makes perfect sense to translate a
subpixel classification of gap fraction (or a linear ratio) into a binary
product.
}
\note{
This method is also available in the HSP software package
\insertCite{Lang2013}{rcaiman}.
}
\examples{
\dontrun{
caim <- read_caim()
r <- caim$Blue
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

path <- system.file("external/ootb_sky.txt", package = "rcaiman")
ootb_sky <- read_ootb_sky_model(gsub(".txt", "", path), z, a)

sky <- ootb_interpolate_and_merge(r, z, a, ootb_sky$sky_points, ootb_sky)

ratio <- r / sky$sky
ratio <- normalize_minmax(ratio, 0, 1, TRUE)
plot(ratio)
g <- sky_grid_segmentation(z, a, 10)
bin2 <- defuzzify(ratio, g)
plot(bin2) # unsatisfactory results due to light conditions
}
}
\references{
\insertAllCited{}
}
\seealso{
Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{calc_spherical_distance}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{display_caim}()},
\code{\link{expand_sky_points}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rel_radiance}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{masking}()},
\code{\link{optim_dist_to_black}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{read_ootb_sky_model}()},
\code{\link{sor_filter}()},
\code{\link{vicinity_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
