% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_caim.R
\name{display_caim}
\alias{display_caim}
\title{Dysplay a canopy image}
\usage{
display_caim(caim = NULL, bin = NULL, g = NULL)
}
\arguments{
\item{caim}{\linkS4class{SpatRaster}, such as the output of \code{\link[=read_caim]{read_caim()}} or
\code{\link[=read_caim_raw]{read_caim_raw()}}, or a single layer raster such as the output of
\code{\link[=enhance_caim]{enhance_caim()}}.}

\item{bin}{description}

\item{g}{description}
}
\value{
No return value. Called for side effects
}
\description{
Wrapper function for \code{\link[EBImage:display]{EBImage::display()}}
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
r <- normalize(caim$Blue)
bin <- regional_thresholding(r, rings_segmentation(z, 30),
                             method = "thr_isodata")
display_caim(caim)
display_caim(bin)
display_caim(c(normalize(caim$Blue), bin))
}
}
\seealso{
Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{calc_sngd}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rel_radiance}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_sky_pixels}()},
\code{\link{masking}()},
\code{\link{optim_dist_to_black}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{read_ootb_sky_model}()},
\code{\link{sor_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
