% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhance_caim.R
\name{enhance_caim}
\alias{enhance_caim}
\title{Enhance canopy image}
\usage{
enhance_caim(
  caim,
  m,
  sky_blue,
  w_red = 0,
  thr = 0.5,
  fuzziness = 10,
  gamma = NULL
)
}
\arguments{
\item{caim}{\linkS4class{SpatRaster}. The return of a call to
\code{\link{read_caim}}.}

\item{m}{\linkS4class{SpatRaster}. A mask. Usually, built with
\code{\link{mask_hs}}.}

\item{sky_blue}{\linkS4class{color}. Is the \code{target_color} argument to
be passed to \code{\link{membership_to_color}}.}

\item{w_red}{Numeric vector of length one. Weight of the red channel. A
single layer image is calculated as a weighted average of the blue and red
channels. This layer is used as lightness information. The weight of the
blue is the complement of \code{w_red}.}

\item{thr}{Numeric vector of length one. Location parameter of the logistic
membership function. Use \code{NULL} to estimate it automatically with the
function \code{\link[autothresholdr]{auto_thresh}}, method
"IsoData".}

\item{fuzziness}{Numeric vector of length one. This number is a constant that
scale \code{mem}. Use \code{NULL} to estimate it automatically as the
midpoint between the maximum and minimum values of \code{lightness}.}

\item{gamma}{Numeric vector of length one. This is for applying a gamma back
correction to the lightness information (see Details and argument
\code{w_red}).}
}
\value{
An object of class \linkS4class{SpatRaster} --with same pixel
  dimensions than \code{caim}-- that should show more contrast between the
  sky and plant pixels than any of the individual bands from \code{caim}.
}
\description{
This function was proposed in \insertCite{Diaz2015;textual}{rcaiman}. It uses
the color perceptual attributes (hue, lightness, and chroma) to enhance the
contrast between the sky and plants through fuzzy classification. The
algorithm was developed following this premise: the color of the sky is
different from the color of plants. It performs the next classification
rules, here expressed in natural language: clear sky is blue and clouds
decrease its chroma; if clouds are highly dense, then the sky is achromatic,
and, in such cases, it can be light or dark; everything that does not match
this description is not sky. These linguistic rules were translated to math
language by means of fuzzy logic.
}
\details{
This is a pixel-wise methodology that evaluates the possibility for a pixel
to be member of the class \emph{Gap}. High score could mean either high
membership to \code{sky_blue} or, in the case of achromatic pixels, a high
membership to values above \code{thr}. The algorithm internally uses
\code{\link{membership_to_color}} and \code{\link{local_fuzzy_thresholding}}.
The argument \code{sky_blue} is the \code{target_color} of the former
function, which output is the argument \code{mem} of the latter function.

The \code{gamma} argument, along with \code{\link{gbc}}, is used to
back-correct the values passed to \code{\link{local_fuzzy_thresholding}}.

If you use this function in your research, please cite
\insertCite{Diaz2015;textual}{rcaiman} in addition to this package.
}
\examples{
\dontrun{
caim <- read_caim()
caim <- normalize(caim, 0, 255)
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
sky_blue_sample <- crop(caim, ext(610,643,760,806))
sky_blue <- apply(sky_blue_sample[], 2, median) \%>\%
  as.numeric() \%>\%
  matrix(., ncol = 3) \%>\%
  sRGB()
ecaim <- enhance_caim(caim, !is.na(z), sky_blue, gamma = 2.2)
plot(ecaim)
}
}
\references{
\insertAllCited{}
}
\seealso{
Other Pre-processing Functions: 
\code{\link{gbc}()},
\code{\link{local_fuzzy_thresholding}()},
\code{\link{membership_to_color}()},
\code{\link{normalize}()}
}
\concept{Pre-processing Functions}
