% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature.R
\name{extract_feature}
\alias{extract_feature}
\title{Extract feature}
\usage{
extract_feature(r, segmentation, fun = mean, return_raster = TRUE)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Single layer raster.}

\item{segmentation}{\linkS4class{RasterLayer}. The segmentation of \code{r}.}

\item{fun}{\code{function} that takes a vector as input and returns a
one-length numeric or logical vector as output (e.g. mean).}

\item{return_raster}{One-length logical vector.}
}
\description{
Extract features from raster images.
}
\details{
Given a single layer raster, a segmentation, and a function,
\code{extract_features} will returns a numeric vector or a
\linkS4class{RasterLayer} depending on whether the parameter
\code{return_raster} is \code{TRUE} or \code{FALSE}. For the first
case, each pixel of each segment will adopt the respective extracted feature
value. For the second case, the return will be the extracted feature as a
vector of length equal to the number of segments in the segmentation. Each
extracted feature value will be obtained by processing all pixels that belong
to a segment with the provided function.
}
\examples{
r <- read_caim()
z <- zenith_image(ncol(r),lens("Nikon_FCE9"))
a <- azimuth_image(z)
g <- sky_grid_segmentation(z, a, 10)
extract_feature(r$Blue, g, return_raster = FALSE)
plot(extract_feature(r$Blue, g, return_raster = FALSE))
}
