% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rel_radiance.R
\name{extract_rel_radiance}
\alias{extract_rel_radiance}
\title{Extract relative radiance}
\usage{
extract_rel_radiance(r, z, a, sky_points, no_of_points = 3, use_window = TRUE)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. The image from which the argument \code{sky_points}
was obtained.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{sky_points}{The output of \code{\link[=extract_sky_points]{extract_sky_points()}}
or an object of the same class and structure.}

\item{no_of_points}{Numeric vector of length one. The number of near-zenith
points required for the estimation of the zenith DN.}

\item{use_window}{Logical vector of length one. If \code{TRUE}, a window of \eqn{3
\times 3} pixels will be used to extract the digital number from \code{r}.}
}
\value{
A list of three objects, \emph{zenith_dn} and \emph{max_zenith_angle} from the
class \emph{numeric}, and \emph{sky_points} of the class
\emph{data.frame}; \emph{zenith_dn} is the estimated zenith digital number,
\emph{max_zenith_angle} is the maximum zenith angle reached in the search
for near-zenith sky points, and \emph{sky_points} is the input argument
\code{sky_points} with the additional columns: \emph{a}, \emph{z},
\emph{dn}, and \emph{rr}, which stand for azimuth and zenith angle in
degrees, digital number, and relative radiance, respectively. If \code{NULL} is
provided as \code{no_of_points}, then \emph{zenith_dn} is forced to one and,
therefore, \emph{dn} and \emph{rr} will be identical.
}
\description{
Extract the radiance relative to the zenith radiance
}
\details{
Interpolate the \code{no_of_points} closer to the zenith using IDW with p = 2.
}
\note{
As an alternative, both \href{https://imagej.net/ij/}{ImageJ} and HSP
software package \insertCite{Lang2013}{rcaiman} can be used to manually
digitize points. See \code{\link[=extract_dn]{extract_dn()}} and \code{\link[=read_manual_input]{read_manual_input()}} for details.
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

# See fit_cie_sky_model() for details on the CSV file
path <- system.file("external/sky_points.csv",
                    package = "rcaiman")
sky_points <- read.csv(path)
sky_points <- sky_points[c("Y", "X")]
colnames(sky_points) <- c("row", "col")
head(sky_points)

plot(caim$Blue)
points(sky_points$col, nrow(caim) - sky_points$row, col = 2, pch = 10)
rr <- extract_rel_radiance(caim$Blue, z, a, sky_points, 1)
points(rr$sky_points$col, nrow(caim) - rr$sky_points$row, col = 3, pch = 0)
}
}
\references{
\insertAllCited{}
}
\seealso{
Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{calc_spherical_distance}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{display_caim}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_sky_pixels}()},
\code{\link{masking}()},
\code{\link{optim_dist_to_black}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{read_ootb_sky_model}()},
\code{\link{sor_filter}()},
\code{\link{vicinity_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
