% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{extract_sky_marks}
\alias{extract_sky_marks}
\title{Extract sky marks}
\usage{
extract_sky_marks(
  r,
  bin,
  z,
  a,
  g,
  path_to_HSP_project,
  img_name,
  dist_to_plant = 3,
  angular_dist = 5,
  min_raster_dist = 9,
  fuzzy_logic = TRUE
)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Greyscale image hemispherical image.}

\item{bin}{\linkS4class{RasterLayer}. The result of binarizing the \code{r}
argument.}

\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}

\item{a}{\code{\linkS4class{RasterLayer}} built with
\code{\link{azimuth_image}}.}

\item{g}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{sky_grid_segmentation}} taking into account the camera, lens,
and pre-processing involved in obtaining the \code{r} argument.}

\item{path_to_HSP_project}{Character vector of length one. Path to the HSP
project folder.}

\item{img_name}{Character vector of length one. The name of the file
associated to the \code{r} argument.}

\item{dist_to_plant}{Numeric vector of length one.}

\item{angular_dist}{Numeric vector of length one. Distance in degrees.}

\item{min_raster_dist}{Numeric vector of length one. Distance in pixels.}

\item{fuzzy_logic}{Logical vector of length one. Default is \code{TRUE}.}
}
\value{
None. A file will be written in the HSP project folder.
}
\description{
Extract sky marks and create a special file to interface with HSP software.
}
\details{
This function is part of a workflow that connects the \code{\link{mblt}}
algorithm with the HSP software package \insertCite{Lang2013}{rcaiman}. This
workflow aims to fully automatize the single camera linear ratio method
presented by \insertCite{Lang2010;textual}{rcaiman}.

The \code{bin} argument should be a binarization from the \code{\link{mblt}}
method set with a high \code{w}. The latter is for obtaining pure sky pixels.
Setting \code{w} to maximum is not recommended because high frequency of
omission errors are expected with that setting.

The pure sky pixels masked by \code{bin} will be automatically sampled by the
function. The density and distribution of the sampling points is controlled
by the arguments \code{g}, \code{dist_to_plant}, \code{angular_dist},
\code{min_raster_dist}, and \code{fuzzy_logic}.

As first step, the maximum pure sky pixel is extracted from every cell of the
sky grid \code{g}. Then, if \code{dist_to_plant} is other than \code{NULL},
the distance to the non-pure sky pixels is computed for every pure sky
pixels, and them are filtrated based on \code{dist_to_plant}. This allows the
user to indicate the minimum gap size involved in the calculations. The
following steps allows to decrease the number of points while maintaining a
good distribution.

If \code{fuzzy_logic} is set to \code{TRUE}, the parameter
\code{angular_dist} is used to compute how much the user considered too close
in the polar space. Then, starting with a given point, fuzzy logic is used to
filter out the darkest and closest points. If \code{fuzzy_logic} is set to
\code{FALSE}, the filter only consider distance and \code{angular_dist}
becomes a threshold.

The \code{min_raster_dist} argument is a threshold and can be used to
indicate the minimum distance in the raster space that is required between
points.
}
\references{
\insertRef{Lang2010}{rcaiman}

\insertRef{Lang2013}{rcaiman}
}
\seealso{
extract_sun_mark
}
