% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{extract_sky_marks}
\alias{extract_sky_marks}
\title{Extract sky marks}
\usage{
extract_sky_marks(r, bin, g, dist_to_plant = 3, min_raster_dist = 3)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Greyscale image hemispherical image.}

\item{bin}{\linkS4class{RasterLayer}. The result of binarizing the \code{r}
argument.}

\item{g}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{sky_grid_segmentation}} taking into account the camera, lens,
and pre-processing involved in obtaining the \code{r} argument.}

\item{dist_to_plant}{Numeric vector of length one.}

\item{min_raster_dist}{Numeric vector of length one. Distance in pixels.}
}
\description{
Extract sky marks for CIE sky model fitting
}
\details{
The \code{bin} argument should be any binarized image that masked out pure
canopy (non-gap) pixels and most of the mixed pixels, so to establish a
region of interest dominated by pure sky pixels (a.k.a., gap pixels). The
\code{bin} argument can be obtained with the \code{\link{ootb_mblt}}.

This function will automatically sample in the sky region delimited by
\code{bin}. The density and distribution of the sampling points is controlled
by the arguments \code{g}, \code{dist_to_plant}, and \code{min_raster_dist}.

As first step, the digital number under the class *Gap* --digital numbers
from \code{r} covered by pixels values equal to one on the \code{bin} layer--
are evaluated to extract its maximum value per cell of the sky grid \code{g}.
But, \code{dist_to_plant} allows users to establish a buffer zone which
modify \code{bin}.

The final step filter these local maximum values.\code{min_raster_dist} is a
minimum distance threshold between points that is applied in the raster
space.
}
\examples{
\dontrun{
require(magrittr)
#' my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
g <- sky_grid_segmentation(z, a, 10)
blue <- gbc(r$Blue)
bin <- ootb_mblt(blue, z, a, w = 0.5)
sky_marks <- extract_sky_marks(blue, bin, g,
                               min_raster_dist = 10)
xy <- cellFromRowCol(z, sky_marks$row, sky_marks$col) \%>\%  xyFromCell(z, .)
plot(blue)
plot(SpatialPoints(xy), add = TRUE, col = 2)
}
}
\concept{hps functions}
