% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sky_points.R
\name{extract_sky_points}
\alias{extract_sky_points}
\title{Extract sky points}
\usage{
extract_sky_points(r, bin, g, dist_to_plant = 3, min_raster_dist = 3)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically,
the blue channel extracted from a hemispherical photograph. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{bin}{\linkS4class{SpatRaster}. This should be a preliminary
binarization of \code{r} useful for masking pixels that are very likely
to be pure sky pixels.}

\item{g}{\linkS4class{SpatRaster} built with
\code{\link{sky_grid_segmentation}}.}

\item{dist_to_plant, min_raster_dist}{Numeric vector of length one or
\code{NULL}.}
}
\value{
An object of the class \emph{data.frame} with two columns named
  \emph{col} and \emph{row}.
}
\description{
Extract sky points for model fitting.
}
\details{
This function will automatically sample sky pixels from the sky region
delimited by \code{bin}. The density and distribution of the sampling points
is controlled by the arguments \code{g}, \code{dist_to_plant}, and
\code{min_raster_dist}.

As the first step, digital numbers from \code{r} covered by pixels values
equal to one on the \code{bin} layer are evaluated to extract its maximum
value per cell of \code{g}. The argument \code{dist_to_plant} allows users to
establish a buffer zone for \code{bin}, meaning a size reduction of areas
with pixels values equal to one.

The final step filters these local maximum values using the
\code{min_raster_dist} argument as a minimum distance threshold between
points that is applied in the raster space.

Using code \code{NULL} as argument skip the filtering step in question.
}
\examples{
\dontrun{
caim <- read_caim() \%>\% normalize()
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
sky_blue_sample <- crop(caim, ext(610,643,760,806))
sky_blue <- apply(sky_blue_sample[], 2, median) \%>\%
  as.numeric() \%>\%
  matrix(., ncol = 3) \%>\%
  sRGB()
ecaim <- enhance_caim(caim, !is.na(z), sky_blue, gamma = 2.2)
bin <- apply_thr(ecaim, 0.75)
g <- sky_grid_segmentation(z, a, 10)
r <- gbc(caim$Blue*255)
sky_points <- extract_sky_points(r, bin, g)
cells <- cellFromRowCol(z, sky_points$row, sky_points$col)
hist(r[cells][,1])
xy <- xyFromCell(z, cells)
plot(r)
plot(vect(xy), add = TRUE, col = 2)
}
}
\seealso{
Other Tools Functions: 
\code{\link{defuzzify}()},
\code{\link{extract_feature}()},
\code{\link{extract_rl}()},
\code{\link{masking}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()}
}
\concept{Tools Functions}
