% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{extract_sun_mark}
\alias{extract_sun_mark}
\title{Extract sun mark}
\usage{
extract_sun_mark(r, bin, z, a, g, max_angular_dist = 45)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Greyscale image hemispherical image.}

\item{bin}{\linkS4class{RasterLayer}. The result of binarizing the \code{r}
argument.}

\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}

\item{a}{\code{\linkS4class{RasterLayer}} built with
\code{\link{azimuth_image}}.}

\item{g}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{sky_grid_segmentation}} taking into account the camera, lens,
and pre-processing involved in obtaining the \code{r} argument.}

\item{max_angular_dist}{Numeric vector of length one. Angle in degree to
establish the maximum size of the sun corona. See details. The former is
raster coordinates of the solar disk (row and column), and the other is
angular coordinates (zenith and azimuth angles in degrees).}
}
\value{
Object of class list with two elements names "row_col" and
  "zenith_azimuth", both are numeric vectors of lenght two.
}
\description{
Extract the sun mark for CIE sky model fitting.
}
\details{
This function use an object-based image analyze theoretical framework. The
segmentation are given by \code{g}, and \code{b} should be understood as an
extra layer. For every cell of \code{g}, the maximum is calculated from the
pixel values on \code{r} that have a value equal to one on \code{bin}. Then,
the quantile 0.95 is extracted from these maximum values and is used to
filter out cells below that threshold, i.e, only the ones with at least one
extremely bright sky pixel is keep.

Selected cells are grouped into segments based on adjacency. The degree of
membership to the class _Sun_ is calculated for every segment by using linear
membership functions for the features brightness --digital number from
\code{r}-- and size -- number of cells that constitute the segment. In other
words, the brighteners and lagers segments are the ones that score higher.
The one with the highest score is selected as a sun seed.

The angular distance from the sun seed to every other segments are computed,
and only the segments not farther than \code{max_angular_dist} are classified
as part of the sun corona. A multi-part segment is created by merging the
sun-corona segments, a convex hull is calculated, and the centroide of its
shape is returned as the sun location in raster coordinates (column and row).

The \code{bin} argument should be the same than for
\code{\link{extract_sky_marks}}.
}
\seealso{
Other hsp functions: 
\code{\link{write_sky_marks}()},
\code{\link{write_sun_mark}()}
}
\concept{hsp functions}
