% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheye_to_pano.R
\name{fisheye_to_pano}
\alias{fisheye_to_pano}
\alias{fisheye_to_pano,RasterStackBrick-method}
\alias{fisheye_to_pano,RasterLayer-method}
\title{Title}
\usage{
fisheye_to_pano(r, g, m = NULL, fill_na = TRUE, fun = mean)

\S4method{fisheye_to_pano}{RasterStackBrick}(r, g, m = NULL, fill_na = TRUE, fun = mean)

\S4method{fisheye_to_pano}{RasterLayer}(r, g, m = NULL, fill_na = TRUE, fun = mean)
}
\arguments{
\item{fun}{}
}
\value{

}
\description{
Title
}
