% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbc.R
\name{gbc}
\alias{gbc}
\title{Gamma back correction}
\usage{
gbc(DN_from_JPEG, gamma = 2.2)
}
\arguments{
\item{DN_from_JPEG}{Numeric vector or object from the
\code{\linkS4class{SpatRaster}} class. Digital numbers from a JPEG file (0
to 255, the standard 8-bit encoded).}

\item{gamma}{Numeric vector of length one. Gamma value. Please see
\insertCite{Diaz2018;textual}{rcaiman} for details.}
}
\value{
The same class than \code{DN_from_JPEG}, with dimension unchanged but
  values rescaled between 0 and 1 in a non linear fashion.
}
\description{
Gamma back correction of JPEG images.
}
\examples{
r <- read_caim()
r
gbc(r)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{normalize}}

Other Pre-processing Functions: 
\code{\link{enhance_caim}()},
\code{\link{local_fuzzy_thresholding}()},
\code{\link{membership_to_color}()},
\code{\link{normalize}()}
}
\concept{Pre-processing Functions}
