% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{gbc}
\alias{gbc}
\title{Gamma back correction}
\usage{
gbc(DN_from_JPEG, gamma = 2.2)
}
\arguments{
\item{DN_from_JPEG}{Numeric vector or an object from the
\code{\linkS4class{Raster}} class. Digital numbers from a JPEG file.}

\item{gamma}{Numeric vector of length one. Gamma value. Please see the
reference for details.}
}
\value{
Normalized \code{\linkS4class{Raster}}.
}
\description{
Gamma back correction
}
\examples{
r <- read_caim()
gbc(r)
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
normalize
}
