% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_and_merge.R
\name{interpolate_and_merge}
\alias{interpolate_and_merge}
\title{Interpolate sky data into a raster and merge it with a sky model raster}
\usage{
interpolate_and_merge(r, z, a, sky_points, ootb_sky, rmax_tune = 1)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. Typically, the blue channel extracted from a
canopy photograph. The image from which \code{sky_points} was obtained.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{sky_points}{An object of class \code{data.frame}. The output of
\code{\link[=extract_sky_points]{extract_sky_points()}} or \link{expand_sky_points}.}

\item{ootb_sky}{An object of the class \code{list} that is the result of calling
\code{\link[=ootb_fit_cie_sky_model]{ootb_fit_cie_sky_model()}}.}

\item{rmax_tune}{Numeric vector of length one. It must be a positive integer.
It is used to fine tune the \code{rmax} argument that is computed internally
(see Details).}
}
\value{
An object of class \linkS4class{SpatRaster}.
}
\description{
This function is part of the efforts to automate the method proposed by
\insertCite{Lang2010;textual}{rcaiman}. A paper for thoroughly presenting and
testing this pipeline is under preparation.
}
\examples{
\dontrun{
caim <- read_caim()
r <- caim$Blue
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

path <- system.file("external/ootb_sky.txt", package = "rcaiman")
ootb_sky <- read_ootb_sky_model(gsub(".txt", "", path))

sky <- interpolate_and_merge(r, z, a, ootb_sky$sky_points[,c("row", "col")],
                             ootb_sky)

bin <- apply_thr(r/sky$sky, 0.75)
plot(bin)

g <- sky_grid_segmentation(z, a, 5, first_ring_different = TRUE)
sky_points <- extract_sky_points(r, bin, g, dist_to_black = 3,
                                 min_raster_dist = 9)

points(ootb_sky$sky_points$col, nrow(caim) - ootb_sky$sky_points$row,
       col = 2, pch = 10)
points(sky_points$col, nrow(caim) - sky_points$row, col = 4, pch = 10)
sky_points <- extract_rel_radiance(r, z, a, sky_points, no_of_points = NULL)$sky_points
i <- sor_filter(sky_points, r, k = 3)

sky2 <- interpolate_and_merge(r, z, a, sky_points[i, c("row", "col")],
                              ootb_sky, rmax_tune = 0.75)
plot(sky2$sky)

sky_points <- expand_sky_points(r, z, a, sky_points[i, c("row", "col")],
                                angle_width = 1.875,
                                k= 3, rmax = 20)
sky_points <- sor_filter(sky_points) \%>\% sky_points[. ,]
sky2 <- interpolate_and_merge(r, z, a, sky_points, ootb_sky, rmax_tune = 0.75)
plot(sky2$sky)
}
}
\references{
\insertAllCited{}
}
\seealso{
Other Sky Reconstruction Functions: 
\code{\link{cie_sky_image}()},
\code{\link{fit_cie_sky_model}()},
\code{\link{fit_coneshaped_model}()},
\code{\link{fit_trend_surface}()},
\code{\link{interpolate_sky_points}()},
\code{\link{ootb_fit_cie_sky_model}()}
}
\concept{Sky Reconstruction Functions}
