% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sky.R
\name{interpolate_dns}
\alias{interpolate_dns}
\title{Build a sky image}
\usage{
interpolate_dns(r, z, a, sky_marks, use_window = TRUE)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}
}
\value{

}
\description{
Build an above canopy image from a single below canopy image.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
sky <- build_sky(blue, z, a)
plot(sky$sky)
}
}
