% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_functions.R
\name{lens}
\alias{lens}
\title{Lens database}
\usage{
lens(type = "equiangular", max_fov = FALSE)
}
\arguments{
\item{type}{Character vector of length one. The name of the lens, see details.}

\item{max_fov}{Logical. Use TRUE to return the maximum FOV in degrees.}
}
\description{
Database of lens projection functions and field of views.
}
\details{
TODO: include the character to retrieve the lens type and the reference
}
\examples{
lens("equiangular")
}
