% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking.R
\name{masking}
\alias{masking}
\title{Image masking}
\usage{
masking(r, m, RGB = c(1, 0, 0))
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. The image. Values should be normalized,
see \code{\link[=normalize]{normalize()}}. Only methods for images with one or three
layers have been implemented.}

\item{m}{\linkS4class{SpatRaster}. A mask. For hemispherical photographs,
check \code{\link[=select_sky_vault_region]{select_sky_vault_region()}}.}

\item{RGB}{Numeric vector of length three. RGB color code. Red is the default
color.}
}
\value{
An object of class \linkS4class{SpatRaster} that is
\code{r} but with areas where \code{m} is equal to zero painted in a solid
color. If \code{r} is a single layer image, then the layer is triplicated
to allow the use of color.
}
\description{
Image masking
}
\examples{
\dontrun{
 r <- read_caim()
 z <- zenith_image(ncol(r), lens())
 a <- azimuth_image(z)
 m <- select_sky_vault_region(z, 20, 70) & select_sky_vault_region(a, 90, 180)

 masked_caim <-  masking(normalize(r), m)
 plotRGB(masked_caim * 255)

 masked_bin <- masking(apply_thr(r$Blue, 125), m)
 plotRGB(masked_bin * 255)
 }

}
\seealso{
\code{\link[=select_sky_vault_region]{select_sky_vault_region()}}

Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{calc_sngd}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{display_caim}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rel_radiance}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_sky_pixels}()},
\code{\link{optim_dist_to_black}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{read_ootb_sky_model}()},
\code{\link{sor_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
