% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblt.R
\name{mblt}
\alias{mblt}
\title{Model-based local thresholding}
\usage{
mblt(r, z, a, intercept = -8, slope = 1, w = 0.5)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}

\item{intercept}{Numeric vector of length one. These are linear
function coefficients. Please, see the Details section of
\code{\link{thr_image}}.}

\item{slope}{Numeric vector of length one. These are linear
function coefficients. Please, see the Details section of
\code{\link{thr_image}}.}

\item{w}{Numeric vector. Weighting parameter for \code{slope}.}
}
\value{
If \code{w} has length equal to one, \linkS4class{RasterLayer}. Else,
  \linkS4class{RasterStack}. If \code{w} is \code{NULL}, the above canopy
  reconstruction is returned.
}
\description{
Out-of-the-box version of the model-based local thresholding (MBLT)
algorithm.
}
\details{
This function is a hard-coded version of a MBLT pipeline that starts with a
working binarized image and ended with a refined binarized image. The
pipeline combines \code{\link{regional_thresholding}},
\code{\link{fit_cone_shaped_model}}, \code{\link{fit_trend_surface_to_sky_dn}}, and
\code{\link{thr_image}}. The code can be easily inspected by calling
\code{mblt}. Advanced users could use that code as a template.

The MBLT algorithm was first presented in
\insertCite{Diaz2018;textual}{rcaiman}. This version differs from that in the
following main aspects:

\itemize{

\item This version implements a regional threholding approach as first step
instead of a global one.

\item It does not use asynchronous acquisition under the open sky.

\item The working binarized image is not refined in the second step by using
the sky digital number (sDN) obtained with Equation 4
\code{\link{fit_cone_shaped_model}} and the local threshold values calculated with
Equation 1 \code{\link{thr_image}}. Instead, the sDN are used as filling
source for trend surface fitting \code{\link{fit_trend_surface_to_sky_dn}}.

\item The sDN obtained by trend surface fitting is merged with the sDN
obtained with Equation 4. To merge them, a weighted average is calculated,
being weights proportional to the zenith angle. Near the zenith, values
obtained by means of trend surface fitting prevail over the ones obtained
with Equation 7, and the opposite occur near the horizon.

}
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)
bin <- mblt(blue, z, a)
}
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
Other mblt functions: 
\code{\link{fit_cone_shaped_model}()},
\code{\link{fit_trend_surface_to_sky_dn}()},
\code{\link{thr_image}()}
}
\concept{mblt functions}
