% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_to_color.R
\name{membership_to_color}
\alias{membership_to_color}
\title{Compute the membership to a target color}
\usage{
membership_to_color(caim, target_color, sigma = NULL)
}
\arguments{
\item{caim}{\linkS4class{SpatRaster}. The output of \code{\link[=read_caim]{read_caim()}}.}

\item{target_color}{\linkS4class{color}.}

\item{sigma}{Numeric vector of length one. Use \code{NULL} (default) to estimate
it automatically as the euclidean distance between \code{target_color} and grey
in the \emph{CIE LAB} color space.}
}
\value{
It returns an object of the class \linkS4class{SpatRaster}. First layer
is the membership to the target color. Second layer is the membership to
grey. Both memberships are calculated with same \code{sigma}.
}
\description{
This function was first presented in \insertCite{Diaz2015;textual}{rcaiman}.
It computes the degree of membership to a color using two Gaussian membership
functions and the axes \emph{A} and \emph{B} from the
\emph{CIE LAB} color space. The lightness dimension is not
considered in the calculations.
}
\details{
If you use this function in your research, please cite
\insertCite{Diaz2015;textual}{rcaiman} in addition to this package
(\verb{citation("rcaiman"}).
}
\examples{
\dontrun{
path <- system.file("external/APC_0581.jpg", package = "rcaiman")
caim <- read_caim(path)
z <- zenith_image(2132/2, lens("Olloclip"))
a <- azimuth_image(z, orientation = 180)
zenith_colrow <- c(1063, 771)/2
caim <- expand_noncircular(caim, z, zenith_colrow)
m <- !is.na(caim$Red) & !is.na(z)
caim[!m] <- 0
caim <- normalize_minmax(caim)

mem <- membership_to_color(caim, sRGB(0.1, 0.4, 0.8))
plot(mem)
}
}
\references{
\insertAllCited{}
}
