% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize,Raster-method}
\alias{normalize,RasterLayer-method}
\title{Normalize data}
\usage{
normalize(r, mn, mx, ...)

\S4method{normalize}{Raster}(r, mn, mx, ...)

\S4method{normalize}{RasterLayer}(r, mn, mx, ...)
}
\arguments{
\item{r}{\code{\linkS4class{Raster}}.}

\item{mn}{Numeric vector of length one. Minimum expected value.}

\item{mx}{Numeric vector of length one. Maximum expected value.}

\item{...}{Additional arguments as for \code{\link[raster]{writeRaster}}.}
}
\description{
Normalize data lying between \code{mn} and \code{mx} in the range \code{0} to
\code{1}. Data greater than \code{mx} get values greater than \code{1} in a
proportional fashion. Conversely, data less than \code{mn} get values less
than \code{0}.
}
\section{Methods (by class)}{
\itemize{
\item \code{Raster}: Result is a \linkS4class{Raster} of the same type that
\code{r}.
}}

\examples{
normalize(read_caim(), 0, 255)

}
