% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obia.R
\name{obia}
\alias{obia}
\title{Object-based image analysis of hemispherical photographs}
\usage{
obia(r, z, a, bin, segmentation, gf_mn = 0.2, gf_mx = 0.95)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photograph. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{SpatRaster}. A working binarized image. This should
be a preliminary binarization of \code{r}.}

\item{segmentation}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{polar_qtree}}.}

\item{gf_mn}{Numeric vector of length one. The minimum gap fraction that a
segment should have to be considered as one containing foliage.}

\item{gf_mx}{Numeric vector of length one. The maximum gap fraction that a
segment should have to be considered as one containing foliage.}
}
\value{
\linkS4class{SpatRaster}. A improved version of \code{bin}.
}
\description{
This method was first presented in \insertCite{Diaz2015;textual}{rcaiman}.
This version is simpler since it relies on a better working binarized image.
The version from 2015 used an automatic selection of samples followed by a
knn classification of segments containing foliage. This version uses de gap
fraction extracted from \code{bin} to classify foliage by defining upper and
lower limits through \code{gf_mx} and \code{gf_mn}.
}
\details{
This method produce a synthetic layer by computing the ratio of \code{r} to
the maximum value of \code{r} at the segment level. This process is carried
out only on the pixels covered by the classes foliage and sky (i.e.,
\code{bin} equal to one). To avoid spurious values, the quantile 0.9 is
computed instead of the maximum. Pixels from the synthetic layer comprised
between 0 an 1 are binarized following two criteria in such a way that to be
turned plant on \code{bin} (i.e, to be assigned as 0) it have to be 0 under
both criteria. Those criteria are \code{defuzzify} with a sky grid
segmentation of 10 degrees and \code{apply_thr} with a threshold equal to
0.5. In addition, it cannot be exclusively surrounded by sky pixels.
}
\examples{
\dontrun{
caim <- read_caim()
caim <- normalize(caim, 0, 255)
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
sky_blue_sample <- crop(caim, ext(610,643,760,806))
sky_blue <- apply(sky_blue_sample[], 2, median) \%>\%
  as.numeric() \%>\%
  matrix(., ncol = 3) \%>\%
  sRGB()
ecaim <- enhance_caim(caim, !is.na(z), sky_blue, gamma = 2.2)
bin <- apply_thr(ecaim, 0.5)
seg <- polar_qtree(caim, z, a)
bin_obia <- obia(gbc(caim$Blue), z, a, bin, seg)
plot(bin - bin_obia)
}
}
\references{
\insertAllCited{}
}
