% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mblt.R
\name{ootb_mblt}
\alias{ootb_mblt}
\title{Out-of-the-box Model-based local thresholding}
\usage{
ootb_mblt(r, z, a)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{RasterLayer}. The result of a call to
\code{\link{azimuth_image}}.}
}
\description{
Out-of-the-box version of the model-based local thresholding (MBLT)
algorithm.
}
\details{
This function is a hard-coded version of a MBLT pipeline that starts with a
working binarized image and ended with a refined binarized image. The
pipeline combines \code{\link{find_sky_dns}},
\code{\link{fit_cone_shaped_model}},
\code{\link{fit_trend_surface_to_sky_dn}}, and \code{\link{thr_image}}. The
code can be easily inspected by calling \code{ootb_mblt} --no parenthesis.
Advanced users could use that code as a template.

The MBLT algorithm was first presented in
\insertCite{Diaz2018;textual}{rcaiman}. This version differs from that in the
following main aspects:

\itemize{

\item $intercept$ is set to 0, $slope$ to 1, and $w$ to 0.5

\item This version implements a regional threholding approach as first step
instead of a global one. Please refer to \code{\link{find_sky_dns}}. The
minimum number of samples (sky dns) required is equals to the 30 % of the
population, considering that it is made of 5 $/times$ 5 sky grid cells.

\item It does not use asynchronous acquisition under the open sky. So, the
conic shaped model (\code{\link{fit_cone_shaped_model}}) run without a
filling source, but the result of it is used as filling source for trend
surface fitting (\code{\link{fit_trend_surface_to_sky_dn}}).

\item The sDN obtained by trend surface fitting is merged with the sDN
obtained with \code{\link{fit_cone_shaped_model}}. To merge them, a weighted
average is calculated, being weights calculated as $zenith angle / 90$ (Near
the zenith, values obtained by means of trend surface fitting prevail over
the ones obtained with the cone shaped model, and the opposite occur near the
horizon).

}
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
a <- azimuth_image(z)
blue <- gbc(r$Blue)

}
}
\references{
\insertRef{Diaz2018}{rcaiman}
}
\seealso{
Other mblt functions: 
\code{\link{find_sky_dns}()},
\code{\link{fit_cone_shaped_model}()},
\code{\link{fit_trend_surface_to_sky_dn}()},
\code{\link{thr_image}()}
}
\concept{mblt functions}
