% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ootb_sky_reconstruction.R
\name{ootb_sky_reconstruction}
\alias{ootb_sky_reconstruction}
\title{Out-of-the-box sky reconstruction}
\usage{
ootb_sky_reconstruction(r, lens_coef)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{lens_coef}{Numeric vector. Polynomial coefficients
of the lens projection function.}
}
\description{
Build an above canopy image from a single below canopy image.
}
\details{
This function is a hard-coded version of a pipeline that uses these main
functions: \code{\link{ootb_mblt}}, \code{\link{fit_cie_sky_model}}, and
\code{\link{interpolate_dns}}. The code can be easily inspected by calling
\code{ootb_sky_reconstruction} --no parenthesis. Advanced users could use
that code as a template.

This pipeline integrate these two studies:
\insertCite{Lang2010;textual}{rcaiman} and
\insertCite{Diaz2018;textual}{rcaiman}.

Details about how the original method by
\insertCite{Diaz2018;textual}{rcaiman} was adapted can be read here
\code{\link{ootb_mblt}}.

The main difference between the original method by
\insertCite{Lang2010;textual}{rcaiman} and the one implemented here are: (1)
it is fully automatic, (2) the interpolation is done in the raster space of a
cylindrical projection instead of the raster space of an hemispherical
equiangular projection, (3) the residuals of the CIE sky model
($residuals=model-data$) are interpolated instead of the sky digital numbers
(the data), and (4) the final sky reconstruction is obtained by subtracting
the interpolated residuals to the CIE sky model instead of by calculating a
weighted average parameterized by the user.

The recommended input for this function is data pre-processed with the HSP
software package \insertCite{Lang2013}{rcaiman}. Please, refer to
\code{\link{write_sky_marks}} for additional details about HSP, and refer to
\code{\link{fit_cie_sky_model}} and \code{\link{interpolate_dns}} to know
why the HSP pre-processing is convenient.
}
\examples{
\dontrun{
my_file <- path.expand("~/DSCN5548.JPG")
download.file("https://osf.io/kp7rx/download", my_file,
              method = "auto", mode = "wb")
r <- read_caim(my_file,
               c(1280, 960) - 745,
               745 * 2,
               745 * 2)
blue <- gbc(r$Blue)
sky <- ootb_sky_reconstruction(blue, lens("Nikon_FCE9"))
plot(sky)
ratio <- blue / sky
plot(ratio)
hist(ratio)
plot(ratio > 1.3)
ratio[ratio > 1.3] <- 1.3
plot(ratio)
}
}
\references{
\insertRef{Lang2010}{rcaiman}

  \insertRef{Lang2013}{rcaiman}

  \insertRef{Diaz2018}{rcaiman}
}
