% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ootb_sky_reconstruction.R
\name{ootb_sky_reconstruction}
\alias{ootb_sky_reconstruction}
\title{Out-of-the-box sky reconstruction}
\usage{
ootb_sky_reconstruction(r, z, a, bin = NULL)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photograph. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{z}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{zenith_image}}.}

\item{a}{\linkS4class{SpatRaster}. The result of a call to
\code{\link{azimuth_image}}.}

\item{bin}{\linkS4class{SpatRaster}. A working binarized image. This should
be a preliminary binarization of \code{r}.}
}
\description{
Build an above canopy image from a single below canopy image.
}
\details{
This function is a hard-coded version of a pipeline that uses these main
functions: \code{\link{ootb_mblt}}, \code{\link{fit_cie_sky_model}}, and
\code{\link{interpolate_sky_points}}. The code can be easily inspected by
calling \code{ootb_sky_reconstruction} --no parenthesis. Advanced users could
use that code as a template.

This pipeline is based on these two studies:
\insertCite{Lang2010;textual}{rcaiman} and
\insertCite{Diaz2018;textual}{rcaiman}.

Details about how the original method by
\insertCite{Diaz2018;textual}{rcaiman} was adapted can be read here
\code{\link{ootb_mblt}}.

The main differences between the original method by
\insertCite{Lang2010;textual}{rcaiman} and the one implemented here are: (1)
it is fully automatic, (2) the residuals of the CIE sky model
($residuals=model-data$) are interpolated instead of the sky digital numbers
(the data), and (3) the final sky reconstruction is obtained by subtracting
the interpolated residuals to the CIE sky model instead of by calculating a
weighted average parameterized by the user.

The recommended input for this function is data pre-processed with the HSP
software package \insertCite{Lang2013}{rcaiman}. Please, refer to
code{link{write_sky_marks}} for additional details about HSP, and refer to
\code{\link{fit_cie_sky_model}} and \code{\link{interpolate_sky_points}} to
know why the HSP pre-processing is convenient.
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
a <- azimuth_image(z)
r <- gbc(caim$Blue)
sky <- ootb_sky_reconstruction(r, z, a)
plot(sky)
ratio <- r / sky
plot(ratio)
hist(ratio)
}
}
\references{
\insertAllCited{}
}
