% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_dist_to_black.R
\name{optim_dist_to_black}
\alias{optim_dist_to_black}
\title{Optimize distance to black}
\usage{
optim_dist_to_black(r, z, a, m, bin, g)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. The image from which the argument \code{sky_points}
was obtained.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{m}{\linkS4class{SpatRaster}. A mask. For hemispherical photographs,
check \code{\link[=select_sky_vault_region]{select_sky_vault_region()}}.}

\item{bin}{\linkS4class{SpatRaster}. This should be a preliminary binarization of
\code{r} useful for masking pixels that are very likely pure sky pixels.}

\item{g}{\linkS4class{SpatRaster} built with \code{\link[=sky_grid_segmentation]{sky_grid_segmentation()}} or
\code{\link[=chessboard]{chessboard()}}.}
}
\value{
Numeric vector of length one.
}
\description{
Optimize distance to black
}
