% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcaiman-package.R
\docType{package}
\name{rcaiman-package}
\alias{rcaiman}
\alias{rcaiman-package}
\title{rcaiman: CAnopy IMage ANalysis}
\description{
Tools for pre-processing and processing canopy photographs with support for raw data reading. Works with images taken with both regular and fisheye lenses (all types). Includes algorithms to mitigate errors caused by non-diffuse light conditions during photograph acquisition.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/GastonMauroDiaz/rcaiman/issues}
}

}
\author{
\strong{Maintainer}: Gastón Mauro Díaz \email{gastonmaurodiaz@gmail.com} (\href{https://orcid.org/0000-0002-0362-8616}{ORCID})

}
\keyword{internal}
