% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcaiman.R
\docType{package}
\name{rcaiman}
\alias{rcaiman}
\title{rcaiman: An R package for CAnopy IMage ANalysis}
\description{
The rcaiman package provides three categories of important functions:
base, diffuse light, and non-difuse. (TODO: revise this categories)
}
\details{
Always err on the side of caution, and simplicity. It’s easier to give people
more functionality than it is to take away stuff they’re used to
https://r-pkgs.org/namespace.html#namespace-workflow
}
\section{Base functions}{

The foo functions ...
}

