% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking_functions.R
\name{read_bin}
\alias{read_bin}
\title{Read binarized images}
\usage{
read_bin(path)
}
\arguments{
\item{path}{One-length character vector. Path to read or a binarized image.}
}
\description{
Wrapper functions for \code{\link[raster]{raster}}.
}
\examples{
\dontrun{
z <- zenith_image(1000, lens())
m <- !is.na(z)
write_bin(m, "mask")
m_from_disk <- read_bin("mask.tif")
plot(m - m_from_disk)
}
}
\seealso{
\code{\link{write_bin}}
}
