% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{read_manual_input}
\alias{read_manual_input}
\title{Read manual input}
\usage{
read_manual_input(path_to_HSP_project, img_name)
}
\arguments{
\item{path_to_HSP_project}{Character vector of length one. Path to the HSP
project folder.}

\item{img_name}{Character vector of length one. See details.}
}
\description{
Read manual input stored in an HSP project.
}
\details{
Refer to the Details section of this function:
\code{\link{write_sky_points}}.
}
\seealso{
Other hps functions: 
\code{\link{read_opt_sky_coef}()},
\code{\link{row_col_from_zenith_azimuth}()},
\code{\link{zenith_azimuth_from_row_col}()}
}
\concept{hps functions}
