% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ootb_sky_model.R
\name{read_ootb_sky_model}
\alias{read_ootb_sky_model}
\title{Read files writen by \code{\link[=write_ootb_sky_model]{write_ootb_sky_model()}}}
\usage{
read_ootb_sky_model(name, z = NULL, a = NULL)
}
\arguments{
\item{name}{Character vector of length one. File name without extension. If
needed, a file path can be added to the names. For example:
\code{"C:/Users/Doe/Documents/DSCN4500"}.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}
}
\value{
An object of the class \emph{list} similar to the output of
\code{\link[=ootb_fit_cie_sky_model]{ootb_fit_cie_sky_model()}}.
}
\description{
Read files writen by \code{\link[=write_ootb_sky_model]{write_ootb_sky_model()}}
}
\examples{
path <- system.file("external/ootb_sky.txt", package = "rcaiman")
ootb_sky <- read_ootb_sky_model(gsub(".txt", "", path))

caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

path <- system.file("external/ootb_sky.txt", package = "rcaiman")
ootb_sky <- read_ootb_sky_model(gsub(".txt", "", path), z, a)
}
\seealso{
Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{display_caim}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rel_radiance}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_sky_pixels}()},
\code{\link{masking}()},
\code{\link{optim_dist_to_black}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{sor_filter}()},
\code{\link{vicinity_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
