% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_thresholding.R
\name{regional_thresholding}
\alias{regional_thresholding}
\title{Regional thresholding}
\usage{
regional_thresholding(r, segmentation, method)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically, the
blue channel extracted from a canopy photograph. Please see \code{\link[=read_caim]{read_caim()}}
and \code{\link[=normalize_minmax]{normalize_minmax()}}.}

\item{segmentation}{\linkS4class{SpatRaster}. The result of segmenting \code{r}.
Arguably, the result of calling \code{\link[=rings_segmentation]{rings_segmentation()}} will be the
preferred choice for fisheye images.}

\item{method}{Character vector of length one. See details for current
options.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and \code{1}.
}
\description{
Regional thresholding of greyscale images.
}
\details{
Methods currently implemented are:
\itemize{
\item \strong{Methods from autothresholdr package}: this function can call
methods from \code{\link[autothresholdr:auto_thresh]{autothresholdr::auto_thresh()}}. For instance, use \code{"IsoData"}
to use the algorithm by \insertCite{isodata;textual}{rcaiman}, which was
recommended by \insertCite{Jonckheere2005;textual}{rcaiman}.
\item \strong{Method isodata from this package}: Use \code{"thr_isodata"} to
use \code{\link[=thr_isodata]{thr_isodata()}}.
\item \strong{Two-corner method from this package}: Use \code{"thr_twocorner"} to
use \code{\link[=thr_twocorner]{thr_twocorner()}} medium threshold.
}
}
\examples{
\dontrun{
path <- system.file("external/DSCN4500.JPG", package = "rcaiman")
zenith_colrow <- c(1276, 980)
diameter <- 756*2
caim <- read_caim(path, zenith_colrow - diameter/2, diameter, diameter)
z <- zenith_image(ncol(caim), lens("Nikon_FCE9"))
r <- gbc(caim$Blue)
r <- correct_vignetting(r, z, c(0.0638, -0.101)) \%>\% normalize_minmax()
rings <- rings_segmentation(z, 15)
bin <- regional_thresholding(r, rings, "thr_isodata")
plot(bin)
}
}
\references{
\insertAllCited{}
}
