% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_thresholding.R
\name{regional_thresholding}
\alias{regional_thresholding}
\title{Regional thresholding}
\usage{
regional_thresholding(
  r,
  segmentation,
  method,
  intercept = NULL,
  slope = NULL,
  prob = NULL
)
}
\arguments{
\item{r}{\linkS4class{SpatRaster}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photograph. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{segmentation}{\linkS4class{SpatRaster}. The result of segmenting
\code{r}. Probably, \code{\link{rings_segmentation}} will be the most used
for fisheye images.}

\item{method}{Character vector of length one. See details for current
options.}

\item{intercept, slope}{Numeric vector of length one. These are linear
function coefficients. Please, see the Details section of
\code{\link{thr_image}}.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and
  \code{1}.
}
\description{
Regional thresholding of greyscale images
}
\details{
Methods currently implemented are:

\itemize{ \item \strong{Diaz2018}: method presented in
\insertCite{Diaz2018;textual}{rcaiman} applied regionally. If this method is
selected, the arguments \code{intercept}, \code{slope}, and \code{prob}
should be provided. It works segmentwise extracting the digital numbers (dns)
per segment and passing them to \code{quantile(dns, prob)}, which aggregated
result (x) is in turn passed to \code{thr_image(x, intercept, slope)}.
Finally, this threshold image is applied to obtain a binarized image.

\item \strong{Methods from autothresholdr package}: this function can call
methods from \code{\link[autothresholdr]{auto_thresh}}. Use \code{"IsoData"}
to use the algorithm by \insertCite{isodata;textual}{rcaiman}, which is the one
recommended by \insertCite{Jonckheere2005;textual}{rcaiman}. }
}
\examples{
\dontrun{
path <- system.file("external/DSCN4500.JPG", package = "rcaiman")
r <- read_caim(path, c(1280, 960) - 745, 745 * 2, 745 * 2)
blue <- gbc(r$Blue)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
rings <- rings_segmentation(z, 10)
bin <- regional_thresholding(blue, rings, "Diaz2018", -8, 0.5, 0.9)
plot(bin) # try ootb_mblt() for better results
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{thr_image}}

Other Tools functions: 
\code{\link{defuzzify}()},
\code{\link{extract_feature}()},
\code{\link{extract_rl}()},
\code{\link{extract_sky_points}()},
\code{\link{gbc}()},
\code{\link{masking}()},
\code{\link{normalize}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()}
}
\concept{Tools functions}
