% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_thresholding.R
\name{regional_thresholding}
\alias{regional_thresholding}
\title{Regional thresholding}
\usage{
regional_thresholding(r, segmentation, method)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. Normalized greyscale image. See
\code{\link{normalize}} and \code{\link{gbc}}}

\item{segmentation}{\linkS4class{RasterLayer}. The result of segmenting
\code{r}. Probably, code{\link{rings_segmentation}} will be the most
used for fisheye images.}

\item{method}{Character vector of length one. See details for current
options.}
}
\value{
\linkS4class{RasterLayer}.
}
\description{
Regional thresholding of greyscale images
}
\details{
Methods currently implemented are:
}
\section{mblt}{
 is the method described in \insertCite{Diaz2018}{rcaiman} but
  applied regionally instead of locally. Digital numbers from each segment
  are extracted and passed to \code{quantile(x, 0.9)}, which result is in
  turn passed to \code{\link{thr_fun}}. The "Generic" coefficients are used.
  Use mblt_default to use \code{w} equal to \code{0.5}, or mblt_initial_value
  to use \code{w} equal to \code{1}.
}

\section{methods from autothresholdr package}{
 this function can call methods
  from \code{\link[autothresholdr]{auto_thresh}}. Use "IsoData" to use the
  algorithm by \insertCite{isodata}{rcaiman}, which is the one recommended by
  \insertCite{Jonckheere2005}{rcaiman}.
}

\examples{
r <- read_caim()
blue <- gbc(r$Blue)
z <- zenith_image(ncol(r), lens("Nikon_FCE9"))
rings <- rings_segmentation(z, 10)
bin <- regional_thresholding(blue, rings, "IsoData")
}
\references{
\insertRef{Diaz2018}{rcaiman}

\insertCite{Jonckheere2005}{rcaiman}

\insertCite{isodata}{rcaiman}
}
