% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_isolated_black_pixels.R
\name{rem_isolated_black_pixels}
\alias{rem_isolated_black_pixels}
\title{Remove isolated black pixels}
\usage{
rem_isolated_black_pixels(bin)
}
\arguments{
\item{bin}{logical \link[terra:SpatRaster-class]{terra::SpatRaster} with a single layer.}
}
\value{
Logical \link[terra:SpatRaster-class]{terra::SpatRaster} of one layer.
}
\description{
Replace single black pixels (\code{FALSE}) that are fully surrounded by white
pixels (\code{TRUE}) with white. Uses 8-connectivity.
}
\examples{
\dontrun{
caim <- read_caim()
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)
r <- caim$Blue

bin <- binarize_by_region(r, ring_segmentation(z, 15), "thr_isodata")

plot(bin)
plot(rem_isolated_black_pixels(bin))
}
}
