% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_isolated_black_pixels.R
\name{rem_isolated_black_pixels}
\alias{rem_isolated_black_pixels}
\title{Remove isolated black pixels}
\usage{
rem_isolated_black_pixels(bin)
}
\arguments{
\item{bin}{\linkS4class{SpatRaster}. Binarized canopy image.}
}
\value{
An object of class \linkS4class{SpatRaster} with values \code{0} and \code{1}.
}
\description{
Remove isolated black pixels
}
\examples{
\dontrun{
caim <- read_caim()
r <- caim$Blue
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)

path <- system.file("external/ootb_sky.txt", package = "rcaiman")
sky <- read_ootb_sky_model(gsub(".txt", "", path), z, a)

bin <- apply_thr(r/sky$sky, 0.9)
plot(bin)
bin2 <- rem_isolated_black_pixels(bin)
plot(bin2)
plot(bin2 - bin)
}
}
\seealso{
Other Binarization Functions: 
\code{\link{apply_thr}()},
\code{\link{obia}()},
\code{\link{ootb_obia}()},
\code{\link{regional_thresholding}()},
\code{\link{thr_isodata}()},
\code{\link{thr_mblt}()}
}
\concept{Binarization Functions}
