% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_col_from_zenith_azimuth.R
\name{row_col_from_zenith_azimuth}
\alias{row_col_from_zenith_azimuth}
\title{Obtain row and col numbers from zenith and azimuth angles}
\usage{
row_col_from_zenith_azimuth(z, a, zenith, azimuth)
}
\arguments{
\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{zenith, azimuth}{Numeric vector. Zenith or azimuth angle in degrees.}
}
\value{
An object of the class \emph{data.frame}.
}
\description{
Obtain row and col numbers from zenith and azimuth angles
}
\examples{
z <- zenith_image(1000, lens())
a <- azimuth_image(z)
row_col_from_zenith_azimuth(z, a, 45, 270)
}
