% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_fisheye_segmentation_functions.R
\name{sky_grid_segmentation}
\alias{sky_grid_segmentation}
\title{Sky grid segmentation}
\usage{
sky_grid_segmentation(z, a, angle_width, sequential = FALSE)
}
\arguments{
\item{z}{\code{\linkS4class{RasterLayer}} built with
\code{\link{zenith_image}}.}

\item{a}{\code{\linkS4class{RasterLayer}} built with
\code{\link{azimuth_image}}.}

\item{angle_width}{One-length numeric vector. It should be \code{30, 15, 10,
7.5, 6, 5, 3.75, 3, 2.5, 1.875, 1} or \code{0.5} degrees. This constrains
is rooted in the requirement of a value able to divide both the \code{0} to
\code{360} and \code{0} to \code{90} ranges into a whole number of
segments.}

\item{sequential}{One-length logical vector. If it is \code{TRUE} the
segments are labeled with sequential numbers. By default (\code{FALSE}),
labeling numbers are not sequential (see Details).}
}
\value{
\linkS4class{RasterLayer} with segments shaped like windshields,
  although some of them will look elongated in height. The pattern is two
  opposite and converging straight sides and two opposite and parallel
  curvy sides.
}
\description{
Segmenting the hemisphere view into segments of equal angular resolution for
both zenith and azimuth angles.
}
\details{
Intersecting rings with sectors makes a grid in which each segment is a
portion of the hemisphere. Each pixel of the grid is labeled with an ID that
codify both ring and sector ID. For example, a grid with a regular interval
of 1 degree has segment from \code{1001} to \code{360090}. This numbers are
calculated with: \code{sectorID * 1000 + ringsID}, where \code{sectorID} is
the ID number of the sector and \code{ringsID} is the ID number of the ring.
}
\examples{
z <- zenith_image(1490, lens())
a <- azimuth_image(z)
g <- sky_grid_segmentation(z, a, 15)
plot(g == 36009)
g <- sky_grid_segmentation(z, a, 15, sequential = TRUE)
plot(g, col = sample(rainbow(length(unique(g)))))
}
\seealso{
Other basic fisheye segmentation functions: 
\code{\link{rings_segmentation}()},
\code{\link{sectors_segmentation}()}
}
\concept{basic fisheye segmentation functions}
