% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sor_filter.R
\name{sor_filter}
\alias{sor_filter}
\title{Statistical outlier removal (SOR) filter}
\usage{
sor_filter(
  sky_points,
  r,
  z,
  a,
  k = 20,
  rmax = 20,
  thr = 2,
  cutoff_side = "both",
  use_window = TRUE
)
}
\arguments{
\item{sky_points}{The output of \code{\link[=extract_sky_points]{extract_sky_points()}}
or an object of the same class and structure.}

\item{r}{\linkS4class{SpatRaster}. The image from which the argument \code{sky_points}
was obtained.}

\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{k}{Numeric vector of length one. Number of k-nearest neighbors.}

\item{rmax}{Numeric vector of length one. The maximum radius for searching
k-nearest neighbors (knn). Points are projected onto a unit-radius sphere,
similar to the use of relative radius in image mapping. The spherical
distance is then calculated and used to filter out points farther than
\code{rmax}.The distance is expressed in degrees. If an insufficient number of
neighbors are found within the search radius, the point is retained due to
a lack of evidence for removal.}

\item{thr}{Numeric vector of length one. See
\insertCite{Leys2013;textual}{rcaiman}.}

\item{cutoff_side}{Character vector of length one. See
\insertCite{Leys2013;textual}{rcaiman}.}

\item{use_window}{Logical vector of length one. If \code{TRUE}, a window of \eqn{3
\times 3} pixels will be used to extract the digital number from \code{r}.}
}
\value{
The argument \code{sky_points} with fewer rows due to outlier removal.
}
\description{
Statistical outlier removal (SOR) filter
}
\details{
This algorithm is based on the homonymous filter from the \href{https://pointclouds.org/}{PCL library}. Distances are computed on a spherical
surface and expressed in degrees to avoid distortions due to projection. The
number of neighbors used for evaluation is controlled by the \code{k} argument, while
the \code{rmax} argument sets the maximum search radius for finding these neighbors.
Points are projected onto a unit-radius sphere, similar to the use of relative
radius in image mapping. The spherical distance is then calculated, and points
farther than rmax are excluded from the neighbor set. If an insufficient
number of neighbors are found within rmax, the point is retained due to a
lack of evidence for removal. The decision criterion follows
\insertCite{Leys2013;textual}{rcaiman}:

\eqn{M - thr \times MAD < x_i < M + thr \times MAD}

where \eqn{x_i} is the value associated with a given sky point, \eqn{M} and
\eqn{MAD} are the median and median absolute deviation, respectively,
computed from the values associated with the neighbors of \eqn{x_i}, and
\eqn{thr} is the user-defined threshold.

The argument \code{cutoff_side} controls which side of the inequality is
tested—either one side or both.
}
\examples{
\dontrun{
caim <- read_caim()
r <- caim$Blue
z <- zenith_image(ncol(caim), lens())
a <- azimuth_image(z)
m <- !is.na(z)
bin <- regional_thresholding(r, rings_segmentation(z, 30),
                             method = "thr_isodata")
bin <- bin & select_sky_vault_region(z, 0, 80)
g <- sky_grid_segmentation(z, a, 5, first_ring_different = TRUE)
sky_points <- extract_sky_points(r, bin, g,
                                 dist_to_black = 3)
plot(r)
points(sky_points$col, nrow(caim) - sky_points$row, col = 2, pch = 10)

sky_points <- sor_filter(sky_points, r, z, a, k = 10, rmax = 20, thr = 2,
                cutoff_side = "left")
points(sky_points$col, nrow(caim) - sky_points$row, col = 3, pch = 0)

}
}
\references{
\insertAllCited{}
}
\seealso{
Other Tool Functions: 
\code{\link{calc_oor_index}()},
\code{\link{calc_spherical_distance}()},
\code{\link{colorfulness}()},
\code{\link{correct_vignetting}()},
\code{\link{defuzzify}()},
\code{\link{display_caim}()},
\code{\link{extract_dn}()},
\code{\link{extract_feature}()},
\code{\link{extract_rel_radiance}()},
\code{\link{extract_sky_points}()},
\code{\link{extract_sun_coord}()},
\code{\link{find_sky_pixels}()},
\code{\link{masking}()},
\code{\link{optim_dist_to_black}()},
\code{\link{optim_normalize}()},
\code{\link{percentage_of_clipped_highlights}()},
\code{\link{read_bin}()},
\code{\link{read_caim}()},
\code{\link{read_caim_raw}()},
\code{\link{read_ootb_sky_model}()},
\code{\link{vicinity_filter}()},
\code{\link{write_bin}()},
\code{\link{write_caim}()},
\code{\link{write_ootb_sky_model}()}
}
\concept{Tool Functions}
