% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_functions.R
\name{test_lens_coef}
\alias{test_lens_coef}
\title{Test lens projection functions}
\usage{
test_lens_coef(lens_coef)
}
\arguments{
\item{lens_coef}{Numeric vector. Polynomial coefficients
of the lens projection function.}
}
\description{
Test that lens projection function works between the 0-to-1 range.
}
\examples{
test_lens_coef(lens("Nikon_FCE9"))
test_lens_coef(pi / 2)
test_lens_coef(c(1.06065, -0.49054, 0.14044))
}
