% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thr_mblt.R
\name{thr_mblt}
\alias{thr_mblt}
\title{Model-based local thresholding}
\usage{
thr_mblt(dn, intercept, slope)
}
\arguments{
\item{dn}{Numeric vector or \linkS4class{SpatRaster}. Digital number of the
background. These values should be normalized and, if they are extracted
from a JPEG image, gamma back corrected.}

\item{intercept, slope}{Numeric vector of length one. These are linear
function coefficients.}
}
\value{
An object of the same class and dimensions than \code{dn}.
}
\description{
Transform background digital number into threshold values
}
\details{
This function transforms background digital numbers into threshold values by
means of the Equation 1 from \insertCite{Diaz2018;textual}{rcaiman}, which is
a linear function with the slope modified by a weighting parameter. This
simple function was found by studying canopy models, also known as targets,
which are perforated surfaces made of a rigid and dark material . These
models were backlighted with homogeneous lighting, photographed with a Nikon
Coolpix 5700 set to acquire in JPEG format, and those images were gamma back
corrected with a default gamma value equal to 2.2 (see \code{\link[=gbc]{gbc()}}). Results
shown that the optimal threshold value was linearly related with the
background digital number (see Figure 1 and Figure 7 from
\insertCite{Diaz2018;textual}{rcaiman}). This shifted the aim from finding
the optimal threshold, following \insertCite{Song2014}{rcaiman} method, to
obtaining the background DN as if the canopy was not there, as
\insertCite{Lang2010}{rcaiman} proposed.
\subsection{Working principle}{

\insertCite{Diaz2018;textual}{rcaiman} observed the following linear
relationship between the background value, usually the sky digital number
(SDN), and the optimal threshold value (OTV):\tabular{llllllllllllr}{
   \eqn{IV = a + b \cdot SDN} \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab (Equation 1a) \cr
}
\tabular{llllllllllllr}{
   \eqn{OTV = a + b \cdot w \cdot SDN} \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab (Equation 1b) \cr
}


were IV is the initial value \insertCite{Wagner2001}{rcaiman}, which is the
boundary between SDN and the mixed pixels, i.e, the pixels that are neither
Gap or Non-gap \insertCite{Macfarlane2011}{rcaiman}, \eqn{a} and \eqn{b} are
the intercept and slope coefficients, respectively, and \eqn{w} is a
weighting parameter that takes into account that OTV is always lower than IV.
If SDN is calculated at the pixel level, a local thresholding method can be
applied by evaluating, pixel by pixel, if the below canopy digital number
(CDN) is greater than the OTV. Formally, If \eqn{CDN>OTV}, then assign \emph{Gap}
class, else assign \emph{Non-gap} class.

This conclusion drawn from an image processing point of view matches with
previous findings drawn from a radiometric measurement paradigm, which are
introduced next.

\insertCite{Cescatti2007;textual}{rcaiman}) posed that cameras can be used as
a radiation measurement device if they are properly calibrated. This method,
denominated by the author as LinearRatio, seeks to obtain the transmittance
(T) as the ratio of below to above canopy radiation:\tabular{llllllllllllr}{
   \eqn{T = CDN/SDN} \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab (Equation 2) \cr
}


were CDN is below canopy digital number (DN), i.e., the DN extracted from a
canopy hemispherical photograph.

The LinearRatio method uses T as a proxy for gap fraction. It ideally
requires twin cameras, one below and the other above the canopy. In contrast,
\insertCite{Lang2010;textual}{rcaiman}) proposed to obtain SDN by manually
selecting pure sky pixels from canopy hemispherical photographs and
reconstructing the whole sky by subsequent modeling and interpolating---this
method is often referred to as LinearRatio single camera or LinearRatioSC.

Equation 2 can be seen as a standardization of the distance between CDN and
SDN. With that in mind, it is useful to rewrite Equation 1b as an inequality
that can be evaluated to return a logical statement that is directly
translated into the desired binary classification:\tabular{llllllllllllr}{
   \eqn{CDN > a + b \cdot w \cdot SDN} \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab (Equation 3) \cr
}


Then, combining Equation 2 and 3, we find that
\insertCite{Diaz2018;textual}{rcaiman} parameters can be applied to T:\tabular{llllllllllllr}{
   \eqn{CDN/SDN > a + b \cdot w \cdot SDN/SDN} \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab (Equation 4a) \cr
}
\tabular{llllllllllllr}{
   \eqn{T > a + b \cdot w} \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab  \tab (Equation 4b) \cr
}


From Equation 2 it is evident that any bias introduced by the camera optical
and electronic system will be canceled during the calculation of T as long as
only one camera is involved. Therefore, After examining Equation 4b, we can
conclude that intercept 0 and slope 1 are theoretically correct.In addition,
the w parameter can be used to filter out mixed pixels and sample the most
pure sky pixels possible. The greater w, the greater the possibility of
selecting pure sky pixels.
}
}
\note{
It is worth noting that Equation 1 was developed with 8-bit images, so
calibration of new coefficient should be done in the 0 to 255 domain since
that is what \code{\link[=thr_mblt]{thr_mblt()}} expect, although the input \code{dn} should be
normalized. The latter was a design decision aiming to harmonize the whole
package, although it might sound counter intuitive.

Nevertheless, new empirical calibrations may be unnecessary since the values
-7.8 \code{intercept} and 0.95 \code{slope} observed with back-gamma corrected JPEG
files produced with the Nikon Coolpix 5700 camera are sufficiently close to
the theoretical values to suggest that using them is a correct assumption. In
addition, instead of calibrating JPEG images, users should adopt raw file
acquisition (\code{\link[=read_caim_raw]{read_caim_raw()}}).

To apply the weighting parameter (w) from Equation 1, just provide the
argument \code{slope} as \eqn{slope \times w}.
}
\examples{
thr_mblt(gbc(125), -8, 1)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=normalize]{normalize()}}, \code{\link[=gbc]{gbc()}}, \code{\link[=apply_thr]{apply_thr()}} and \code{\link[=regional_thresholding]{regional_thresholding()}}.

Other Binarization Functions: 
\code{\link{apply_thr}()},
\code{\link{find_sky_pixels_nonnull}()},
\code{\link{find_sky_pixels}()},
\code{\link{obia}()},
\code{\link{ootb_mblt}()},
\code{\link{ootb_obia}()},
\code{\link{regional_thresholding}()},
\code{\link{thr_isodata}()}
}
\concept{Binarization Functions}
