% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masking_functions.R
\name{write_bin}
\alias{write_bin}
\title{Write binarized images}
\usage{
write_bin(bin, path)
}
\arguments{
\item{bin}{\linkS4class{RasterLayer}.}

\item{path}{Character vector of length one. Path for writing the image.}
}
\description{
Wrapper functions for \code{\link[raster]{writeRaster}}.
}
\examples{
\dontrun{
z <- zenith_image(1000, lens())
m <- !is.na(z)
write_bin(m, "mask")
m_from_disk <- read_bin("mask.tif")
plot(m - m_from_disk)
}
}
\seealso{
\code{\link{read_bin}}
}
