% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_caim.R
\name{write_caim}
\alias{write_caim}
\title{Write canopy image}
\usage{
write_caim(caim, path, bit_depth)
}
\arguments{
\item{caim}{\linkS4class{Raster}.}

\item{path}{One-length character vector. Path for writing the image.}

\item{bit_depth}{One-length numeric vector.}
}
\description{
Wrapper function for \code{\link[raster]{writeRaster}}.
}
\examples{
\dontrun{
caim <- read_caim()
ncaim <- normalize(caim, 0, 255)
write_caim(caim * 2^8, "test_8bit", 8)
write_caim(caim * 2^16, "test_16bit", 16)
}
}
\seealso{
write_bin
}
