% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hps_functions.R
\name{zenith_azimuth_from_row_col}
\alias{zenith_azimuth_from_row_col}
\title{Zenith and azimuth angles from row and col numbers}
\usage{
zenith_azimuth_from_row_col(r, row_col, lens_coef)
}
\arguments{
\item{r}{\linkS4class{RasterLayer}. A normalized greyscale image. Typically,
the blue channel extracted from an hemispherical photographs. Please see
\code{\link{read_caim}} and \code{\link{normalize}}.}

\item{row_col}{Numeric vector of length two. Row and col numbers.}
}
\description{
Zenith and azimuth angles from row and col numbers
}
\examples{
z <- zenith_image(1000, lens_coef = lens())
zenith_azimuth_from_row_col(z, c(501, 751), lens())
}
\seealso{
Other hps functions: 
\code{\link{extract_sky_marks}()},
\code{\link{read_manual_input}()},
\code{\link{read_opt_sky_coef}()},
\code{\link{row_col_from_zenith_azimuth}()}
}
\concept{hps functions}
