% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_azimuth_from_row_col.R
\name{zenith_azimuth_from_row_col}
\alias{zenith_azimuth_from_row_col}
\title{Obtain zenith and azimuth angles from row and col numbers}
\usage{
zenith_azimuth_from_row_col(z, a, row, col)
}
\arguments{
\item{z}{\linkS4class{SpatRaster} built with \code{\link[=zenith_image]{zenith_image()}}.}

\item{a}{\linkS4class{SpatRaster} built with \code{\link[=azimuth_image]{azimuth_image()}}.}

\item{row, col}{Numeric vector. Row or column numbers.}
}
\value{
An object of the class \emph{data.frame}.
}
\description{
Obtain zenith and azimuth angles from row and col numbers
}
\examples{
z <- zenith_image(1000, lens())
a <- azimuth_image(z)
zenith_azimuth_from_row_col(z, a, 499, 751)
}
