% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith_image.R
\name{zenith_image}
\alias{zenith_image}
\title{Zenith image}
\usage{
zenith_image(diameter, lens_coef)
}
\arguments{
\item{diameter}{Numeric vector of length one. Diameter in pixels.}

\item{lens_coef}{Numeric vector. Polynomial coefficients
of the lens projection function.}
}
\value{
\code{\linkS4class{RasterLayer}}.
}
\description{
Built a single layer image with zenith angles values.
}
\examples{
zenith_image(1490, lens("Nikon_FCE9"))
}
